@@ -1,12 +1,11 @@
 {
-    "name": "metafox/scribe",
+    "name": "knuckleswtf/scribe",
     "license": "MIT",
     "description": "Generate API documentation for humans from your Laravel codebase.✍",
     "keywords": [
         "API",
         "documentation",
-        "laravel",
-        "dingo"
+        "laravel"
     ],
     "homepage": "http://github.com/knuckleswtf/scribe",
     "authors": [
@@ -15,45 +14,45 @@
         }
     ],
     "require": {
-        "php": ">=8.0",
+        "php": ">=8.1",
         "ext-fileinfo": "*",
         "ext-json": "*",
         "ext-pdo": "*",
         "erusev/parsedown": "1.7.4",
-        "fakerphp/faker": "^1.9.1",
-        "illuminate/console": "^8.0|^9.0|^10.0",
-        "illuminate/routing": "^8.0|^9.0|^10.0",
-        "illuminate/support": "^8.0|^9.0|^10.0",
-        "league/flysystem": "^1.1.4|^2.1.1|^3.0",
+        "fakerphp/faker": "^1.23.1",
+        "laravel/framework": "^9.0|^10.0|^11.0|^12.0",
+        "league/flysystem": "^3.0",
         "mpociot/reflection-docblock": "^1.0.1",
-        "nikic/php-parser": "^4.10",
-        "nunomaduro/collision": "^5.10|^6.0|^7.0",
+        "nikic/php-parser": "^5.0",
+        "nunomaduro/collision": "^6.0|^7.0|^8.0",
         "ramsey/uuid": "^4.2.2",
         "shalvah/clara": "^3.1.0",
-        "shalvah/upgrader": "^0.3.0",
+        "shalvah/upgrader": ">=0.6.0",
         "spatie/data-transfer-object": "^2.6|^3.0",
-        "symfony/var-exporter": "^5.4|^6.0",
-        "symfony/yaml": "^5.4|^6.0"
+        "symfony/var-exporter": "^6.0|^7.0",
+        "symfony/yaml": "^6.0|^7.0"
     },
     "require-dev": {
-        "brianium/paratest": "^6.0",
-        "dms/phpunit-arraysubset-asserts": "^0.4",
+        "dms/phpunit-arraysubset-asserts": "^v0.5.0",
         "laravel/legacy-factories": "^1.3.0",
-        "laravel/lumen-framework": "^8.0|^9.0|^10.0",
         "league/fractal": "^0.20",
         "nikic/fast-route": "^1.3",
-        "orchestra/testbench": "^6.0|^7.0|^8.0",
-        "pestphp/pest": "^1.21",
-        "phpstan/phpstan": "^1.0",
-        "phpunit/phpunit": "^9.0|^10.0",
-        "symfony/css-selector": "^5.4|^6.0",
-        "symfony/dom-crawler": "^5.4|^6.0"
+        "orchestra/testbench": "^7.0|^8.0|^v9.10.0|^10.0",
+        "pestphp/pest": "^1.21|^2.0|^3.0",
+        "phpstan/phpstan": "^2.1.5",
+        "phpunit/phpunit": "^9.0|^10.0|^11.0",
+        "symfony/css-selector": "^6.0|^7.0",
+        "symfony/dom-crawler": "^6.0|^7.0",
+        "spatie/ray": "^1.41"
     },
     "autoload": {
         "psr-4": {
             "Knuckles\\Scribe\\": "src/",
             "Knuckles\\Camel\\": "camel/"
-        }
+        },
+        "files": [
+          "src/Config/helpers.php"
+        ]
     },
     "autoload-dev": {
         "psr-4": {
@@ -62,10 +61,8 @@
     },
     "scripts": {
         "lint": "phpstan analyse -c ./phpstan.neon src camel --memory-limit 1G",
-        "test": "pest --stop-on-failure --exclude-group dingo --colors",
-        "test-ci": "pest --exclude-group dingo --coverage --min=80",
-        "test-parallel": "paratest -p16 --stop-on-failure --exclude-group dingo",
-        "test-parallel-ci": "paratest -p16 --exclude-group dingo"
+        "test": "pest --stop-on-failure --colors",
+        "test-ci": "pest --coverage --min=80"
     },
     "extra": {
         "laravel": {
