@@ -2,6 +2,7 @@
 
 namespace Knuckles\Scribe\Extracting;
 
+use Closure;
 use Illuminate\Contracts\Validation\Rule;
 use Illuminate\Contracts\Validation\ValidationRule;
 use Illuminate\Support\Arr;
@@ -14,6 +15,7 @@ use Knuckles\Scribe\Exceptions\ScribeException;
 use Knuckles\Scribe\Tools\ConsoleOutputUtils as c;
 use Knuckles\Scribe\Tools\WritingUtils as w;
 use ReflectionClass;
+use ReflectionFunction;
 use Throwable;
 
 trait ParsesValidationRules
@@ -22,22 +24,20 @@ trait ParsesValidationRules
 
     public static \stdClass $MISSING_VALUE;
 
-    public function getParametersFromValidationRules(array $validationRules, array $customParameterData = []): array
+    public function getParametersFromValidationRules(array $validationRulesByParameters, array $customParameterData = []): array
     {
         self::$MISSING_VALUE = new \stdClass();
-        $validationRules = $this->normaliseRules($validationRules);
+        $validationRulesByParameters = $this->normaliseRules($validationRulesByParameters);
 
         $parameters = [];
-        $dependentRules = [];
-        foreach ($validationRules as $parameter => $ruleset) {
+        $rulesWhichDependOnType = ['between', 'max', 'min', 'size', 'gt', 'gte', 'lt', 'lte', 'before', 'after', 'before_or_equal', 'after_or_equal'];
+        foreach ($validationRulesByParameters as $parameter => $ruleset) {
+            $userSpecifiedParameterInfo = $customParameterData[$parameter] ?? [];
+            $stringRules = array_filter($ruleset, fn($rule) => is_string($rule));
+            $rulesAndArguments = array_map(fn($rule) => $this->parseStringRuleIntoRuleAndArguments($rule), $stringRules);
+
             try {
-                $parameterPlusDot = $parameter . '.';
-                if (count($customParameterData) && !isset($customParameterData[$parameter])
-                    && ! Arr::first(array_keys($customParameterData), fn ($key) => str_starts_with($key, $parameterPlusDot))
-                ) {
-                    c::debug($this->getMissingCustomDataMessage($parameter));
-                }
-                $userSpecifiedParameterInfo = $customParameterData[$parameter] ?? [];
+                $this->warnAboutMissingCustomParameterData($parameter, $customParameterData);
 
                 // Make sure the user-specified description comes first (and add full stops where needed).
                 $description = $userSpecifiedParameterInfo['description'] ?? '';
@@ -50,47 +50,67 @@ trait ParsesValidationRules
                     'type' => null,
                     'example' => self::$MISSING_VALUE,
                     'description' => $description,
+                    'nullable' => false,
                 ];
-                $dependentRules[$parameter] = [];
 
-                // First, parse only "independent" rules
-                foreach ($ruleset as $rule) {
-                    $parsed = $this->parseRule($rule, $parameterData, true);
-                    if (!$parsed) {
-                        $dependentRules[$parameter][] = $rule;
-                    }
+                $closureRules = array_filter($ruleset, fn($rule) => ($rule instanceof ClosureValidationRule || $rule instanceof Closure));
+                foreach ($closureRules as $rule) {
+                    $this->processClosureRule($rule, $parameterData);
                 }
 
-                $parameterData['description'] = trim($parameterData['description']);
+                $enumValidationRules = array_filter($ruleset, fn($rule) => $rule instanceof \Illuminate\Validation\Rules\Enum);
+                foreach ($enumValidationRules as $rule) {
+                    $this->processEnumValidationRule($rule, $parameterData);
+                }
 
-                // Set a default type
+                $ruleObjects = array_filter($ruleset, fn($rule) => ($rule instanceof Rule || $rule instanceof ValidationRule));
+                foreach ($ruleObjects as $rule) {
+                    $this->processRuleObject($rule, $parameterData);
+                }
+
+                // TODO support more rules
+                // Process in 3 passes
+                // 1. Rules which provide no info about type or example
+                // (required, required_*, same, different, nullable, exists, and others in the "utilities" group)
+                // 2. Rules which set a type.
+                // 3. Rules whose processing depends on the type. ('between', 'max', 'min', 'size', 'gt', 'gte', 'lt', 'lte')
+                // - Note: 'in' does not provide type info (does it?), but is enough to generate an example
+
+                // First pass: process rules which provide no type or example info
+                $firstPassRuleNames = [
+                    "required",
+                    "required_*",
+                    "same",
+                    "different",
+                    "nullable",
+                ];
+
+                $firstPassRules = array_filter($rulesAndArguments, fn($ruleAndArgs) => Str::is($firstPassRuleNames, $ruleAndArgs[0]));
+                foreach ($firstPassRules as $ruleAndArgs) {
+                    $this->processRule($ruleAndArgs[0], $ruleAndArgs[1], $parameterData, $validationRulesByParameters);
+                }
+
+                $secondPassRules = array_filter($rulesAndArguments, fn($ruleAndArgs) => !Str::is($firstPassRuleNames, $ruleAndArgs[0]) && !in_array($ruleAndArgs[0], $rulesWhichDependOnType));
+                foreach ($secondPassRules as $ruleAndArgs) {
+                    $this->processRule($ruleAndArgs[0], $ruleAndArgs[1], $parameterData, $validationRulesByParameters);
+                }
+
+                // The second pass should have set a type. If not, set a default type
                 if (is_null($parameterData['type'])) {
                     $parameterData['type'] = 'string';
                 }
 
-                $parameterData['name'] = $parameter;
-                $parameters[$parameter] = $parameterData;
-            } catch (Throwable $e) {
-                if ($e instanceof ScribeException) {
-                    // This is a lower-level error that we've encountered and wrapped;
-                    // Pass it on to the user.
-                    throw $e;
-                }
-                throw ProblemParsingValidationRules::forParam($parameter, $e);
-            }
-        }
-
-        // Now parse any "dependent" rules and set examples. At this point, we should know all field's types.
-        foreach ($dependentRules as $parameter => $ruleset) {
-            try {
-                $parameterData = $parameters[$parameter];
-                $userSpecifiedParameterInfo = $customParameterData[$parameter] ?? [];
-
-                foreach ($ruleset as $rule) {
-                    $this->parseRule($rule, $parameterData, false, $parameters);
+                if ($parameterData['required'] === true) {
+                    $parameterData['nullable'] = false;
                 }
 
-                // Make sure the user-specified example overwrites others.
+                // Now parse any "dependent" rules and set examples. At this point, we should know all field's types.
+                $thirdPassRules = array_filter($rulesAndArguments, fn($ruleAndArgs) => in_array($ruleAndArgs[0], $rulesWhichDependOnType));
+                foreach ($thirdPassRules as $ruleAndArgs) {
+                    $this->processRule($ruleAndArgs[0], $ruleAndArgs[1], $parameterData, $validationRulesByParameters);
+                }
+
+                // Make sure the user-specified example overwrites ours.
                 if (array_key_exists('example', $userSpecifiedParameterInfo)) {
                     if ($userSpecifiedParameterInfo['example'] != null && $this->shouldCastUserExample()) {
                         // Examples in comments are strings, we need to cast them properly
@@ -101,11 +121,11 @@ trait ParsesValidationRules
                 }
 
                 // End descriptions with a full stop
-                $parameterData['description'] = trim($parameterData['description']);
                 if (!empty($parameterData['description']) && !Str::endsWith($parameterData['description'], '.')) {
                     $parameterData['description'] .= '.';
                 }
 
+                $parameterData['description'] = trim($parameterData['description']);
                 $parameters[$parameter] = $parameterData;
             } catch (Throwable $e) {
                 if ($e instanceof ScribeException) {
@@ -121,8 +141,10 @@ trait ParsesValidationRules
     }
 
     /**
-     * Transform validation rules from:
+     * Transform validation rules:
+     * 1. from strings to arrays:
      * 'param1' => 'int|required'  TO  'param1' => ['int', 'required']
+     * 2. from '*.foo' to '
      *
      * @param array<string,string|string[]> $rules
      *
@@ -146,8 +168,16 @@ trait ParsesValidationRules
         // Nested array parameters will be present, with '*' replaced by '0'
         $newRules = Validator::make($testData, $rules)->getRules();
 
-        // Transform the key names back from 'ids.0' to 'ids.*'
         return collect($newRules)->mapWithKeys(function ($val, $paramName) use ($rules) {
+            // Transform the key names back from '__asterisk__' to '*'
+            if (Str::contains($paramName, '__asterisk__')) {
+                // In Laravel < v11.44, * keys were replaced with only "__asterisk__"
+                // After that, * keys were replaced with "__asterisk__<random placeholder>", eg "__asterisk__dkjiu78gujjhb
+                // See https://github.com/laravel/framework/pull/54845/
+                $paramName = preg_replace('/__asterisk__[^.]*\b/', '*', $paramName);
+            }
+
+            // Transform the key names back from 'ids.0' to 'ids.*'
             if (Str::contains($paramName, '.0')) {
                 $genericArrayKeyName = str_replace('.0', '.*', $paramName);
 
@@ -161,96 +191,88 @@ trait ParsesValidationRules
         })->toArray();
     }
 
+    // For inline Closure rules, turn any comment above it into a description
+    //
+    // $request->validate([
+    // 'my_param' => [
+    // 'required',
+    //     /** Must be a hexadecimal number. */
+    //     function ($attribute, $value, $fail) {
+    //         if (!preg_match('/^[0-9a-f]+$/', $value)) {
+    //             $fail('Must be in hex format');
+    //         }
+    //     },
+    // ],
+    // ]);
+    protected function processClosureRule($rule, array &$parameterData): void
+    {
+        $docComment = (new ReflectionFunction($rule instanceof ClosureValidationRule ? $rule->callback : $rule))
+            ->getDocComment();
+
+        if (is_string($docComment)) {
+            $description = '';
+            foreach (explode("\n", $docComment) as $line) {
+                $cleaned = preg_replace(['/\*+\/$/', '/^\/\*+\s*/', '/^\*+\s*/'], '', trim($line));
+                if ($cleaned != '') $description .= ' ' . $cleaned;
+            }
+
+            $parameterData['description'] .= $description;
+        }
+    }
+
+    protected function processEnumValidationRule($rule, array &$parameterData, array $allParameters = []): void
+    {
+        $property = (new ReflectionClass($rule))->getProperty('type');
+        $property->setAccessible(true);
+        $enumClass = $property->getValue($rule);
+
+        if (enum_exists($enumClass) && method_exists($enumClass, 'tryFrom')) {
+            // $case->value only exists on BackedEnums, not UnitEnums
+            // method_exists($enum, 'tryFrom') implies the enum is a BackedEnum
+            // @phpstan-ignore-next-line
+            $cases = array_map(fn($case) => $case->value, $enumClass::cases());
+            $parameterData['type'] = gettype($cases[0]);
+            $parameterData['enumValues'] = $cases;
+            $parameterData['setter'] = fn() => Arr::random($cases);
+        }
+    }
+
+    protected function processRuleObject($rule, array &$parameterData): void
+    {
+        if (method_exists($rule, 'invokable')) {
+            // Laravel wraps InvokableRule instances in an InvokableValidationRule class,
+            // so we must retrieve the original rule
+            $rule = $rule->invokable();
+        }
+
+        // Users can define a custom "docs" method on a rule to give Scribe more info.
+        if (method_exists($rule, 'docs')) {
+            $customData = call_user_func_array([$rule, 'docs'], []) ?: [];
+
+            if (isset($customData['description'])) {
+                $parameterData['description'] .= ' ' . $customData['description'];
+            }
+            if (isset($customData['example'])) {
+                $parameterData['setter'] = fn() => $customData['example'];
+            } elseif (isset($customData['setter'])) {
+                $parameterData['setter'] = $customData['setter'];
+            }
+
+            $parameterData = array_merge($parameterData, Arr::except($customData, [
+                'description', 'example', 'setter',
+            ]));
+        }
+    }
+
     /**
      * Parse a validation rule and extract a parameter type, description and setter (used to generate an example).
-     *
-     * If $independentOnly is true, only independent rules will be parsed.
-     * If a rule depends on another parameter (eg gt:field) or needs to know the type of the parameter first (eg:
-     * size:34), it will return false.
+     * @param array $allParameters All parameters, used to check if an argument in the date rules (eg `before:some_date`)
+     *   is a parameter in the request body.
      */
-    protected function parseRule($rule, array &$parameterData, bool $independentOnly, array $allParameters = []): bool
+    protected function processRule($rule, $ruleArguments, array &$parameterData, array $allParameters = []): bool
     {
-        // Reminders:
-        // 1. Append to the description (with a leading space); don't overwrite.
-        // 2. Avoid testing on the value of $parameterData['type'],
-        // as that may not have been set yet, since the rules can be in any order.
-        // For this reason, only deterministic rules are supported
-        // 3. All rules supported must be rules that we can generate a valid dummy value for.
-
-        if ($rule instanceof ClosureValidationRule || $rule instanceof \Closure) {
-            $reflection = new \ReflectionFunction($rule instanceof ClosureValidationRule ? $rule->callback : $rule);
-
-            if (is_string($description = $reflection->getDocComment())) {
-                $finalDescription = '';
-                // Cleanup comment block and extract just the description
-                foreach (explode("\n", $description) as $line) {
-                    $cleaned = preg_replace(['/\*+\/$/', '/^\/\*+\s*/', '/^\*+\s*/'], '', trim($line));
-                    if ($cleaned != '') $finalDescription .= ' ' . $cleaned;
-                }
-
-                $parameterData['description'] .= $finalDescription;
-            }
-
-            return true;
-        }
-
-        if (function_exists('enum_exists') && $rule instanceof \Illuminate\Validation\Rules\Enum) {
-            $reflection = new \ReflectionClass($rule);
-            $property = $reflection->getProperty('type');
-            $property->setAccessible(true);
-            $type = $property->getValue($rule);
-
-            if (enum_exists($type) && method_exists($type, 'tryFrom')) {
-                $cases = array_map(fn ($case) => $case->value, $type::cases());
-                $parameterData['type'] = gettype($cases[0]);
-                $parameterData['description'] .= ' Must be one of ' . w::getListOfValuesAsFriendlyHtmlString($cases) . ' ';
-                $parameterData['setter'] = fn () => Arr::random($cases);
-            }
-
-            return true;
-        }
-
-        if ($rule instanceof Rule || $rule instanceof ValidationRule) {
-            if (method_exists($rule, 'invokable')) {
-                // Laravel wraps InvokableRule instances in an InvokableValidationRule class,
-                // so we must retrieve the original rule
-                $rule = $rule->invokable();
-            }
-
-            if (method_exists($rule, 'docs')) {
-                $customData = call_user_func_array([$rule, 'docs'], []) ?: [];
-
-                if (isset($customData['description'])) {
-                    $parameterData['description'] .= ' ' . $customData['description'];
-                }
-                if (isset($customData['example'])) {
-                    $parameterData['setter'] = fn() => $customData['example'];
-                } elseif (isset($customData['setter'])) {
-                    $parameterData['setter'] = $customData['setter'];
-                }
-
-                $parameterData = array_merge($parameterData, Arr::except($customData, [
-                    'description', 'example', 'setter',
-                ]));
-            }
-
-            return true;
-        }
-
-        if (!is_string($rule)) {
-            return false;
-        }
-
+        // Reminder: Always append to the description (with a leading space); don't overwrite.
         try {
-            // Convert string rules into rule + arguments (eg "in:1,2" becomes ["in", ["1", "2"]])
-            $parsedRule = $this->parseStringRuleIntoRuleAndArguments($rule);
-            [$rule, $arguments] = $parsedRule;
-
-            $dependentRules = ['between', 'max', 'min', 'size', 'gt', 'gte', 'lt', 'lte', 'before', 'after', 'before_or_equal', 'after_or_equal'];
-            if ($independentOnly && in_array($rule, $dependentRules)) {
-                return false;
-            }
-
             switch ($rule) {
                 case 'required':
                     $parameterData['required'] = true;
@@ -259,6 +281,11 @@ trait ParsesValidationRules
                     $parameterData['required'] = true;
                     $parameterData['type'] = 'boolean';
                     $parameterData['description'] .= ' Must be accepted.';
+                    $parameterData['setter'] = fn() => true;
+                    break;
+                case 'accepted_if':
+                    $parameterData['type'] = 'boolean';
+                    $parameterData['description'] .= " Must be accepted when <code>$ruleArguments[0]</code> is " . w::getListOfValuesAsFriendlyHtmlString(array_slice($ruleArguments, 1));
                     $parameterData['setter'] = fn() => true;
                     break;
 
@@ -364,18 +391,18 @@ trait ParsesValidationRules
                 case 'date_format':
                     $parameterData['type'] = 'string';
                     // Laravel description here is "The value must match the format Y-m-d". Not descriptive enough.
-                    $parameterData['description'] .= " Must be a valid date in the format <code>{$arguments[0]}</code>.";
-                    $parameterData['setter'] = function () use ($arguments) {
-                        return date($arguments[0], time());
+                    $parameterData['description'] .= " Must be a valid date in the format <code>{$ruleArguments[0]}</code>.";
+                    $parameterData['setter'] = function () use ($ruleArguments) {
+                        return date($ruleArguments[0], time());
                     };
                     break;
                 case 'after':
                 case 'after_or_equal':
                     $parameterData['type'] = 'string';
-                    $parameterData['description'] .= ' ' . $this->getDescription($rule, [':date' => "<code>{$arguments[0]}</code>"]);
+                    $parameterData['description'] .= ' ' . $this->getDescription($rule, [':date' => "<code>{$ruleArguments[0]}</code>"]);
                     // TODO introduce the concept of "modifiers", like date_format
                     // The startDate may refer to another field, in which case, we just ignore it for now.
-                    $startDate = isset($allParameters[$arguments[0]]) ? 'today' : $arguments[0];
+                    $startDate =  array_key_exists($ruleArguments[0], $allParameters) ? 'today' : $ruleArguments[0];
                     $parameterData['setter'] = fn() => $this->getFaker()->dateTimeBetween($startDate, '+100 years')->format('Y-m-d');
                     break;
                 case 'before':
@@ -383,38 +410,38 @@ trait ParsesValidationRules
                     $parameterData['type'] = 'string';
                     // The argument can be either another field or a date
                     // The endDate may refer to another field, in which case, we just ignore it for now.
-                    $endDate = isset($allParameters[$arguments[0]]) ? 'today' : $arguments[0];
-                    $parameterData['description'] .= ' ' . $this->getDescription($rule, [':date' => "<code>{$arguments[0]}</code>"]);
+                    $endDate = array_key_exists($ruleArguments[0], $allParameters) ? 'today' : $ruleArguments[0];
+                    $parameterData['description'] .= ' ' . $this->getDescription($rule, [':date' => "<code>{$ruleArguments[0]}</code>"]);
                     $parameterData['setter'] = fn() => $this->getFaker()->dateTimeBetween('-30 years', $endDate)->format('Y-m-d');
                     break;
                 case 'starts_with':
-                    $parameterData['description'] .= ' Must start with one of ' . w::getListOfValuesAsFriendlyHtmlString($arguments);
-                    $parameterData['setter'] = fn() => $this->getFaker()->lexify("{$arguments[0]}????");;
+                    $parameterData['description'] .= ' Must start with one of ' . w::getListOfValuesAsFriendlyHtmlString($ruleArguments);
+                    $parameterData['setter'] = fn() => $this->getFaker()->lexify("{$ruleArguments[0]}????");;
                     break;
                 case 'ends_with':
-                    $parameterData['description'] .= ' Must end with one of ' . w::getListOfValuesAsFriendlyHtmlString($arguments);
-                    $parameterData['setter'] = fn() => $this->getFaker()->lexify("????{$arguments[0]}");;
+                    $parameterData['description'] .= ' Must end with one of ' . w::getListOfValuesAsFriendlyHtmlString($ruleArguments);
+                    $parameterData['setter'] = fn() => $this->getFaker()->lexify("????{$ruleArguments[0]}");;
                     break;
                 case 'uuid':
                     $parameterData['description'] .= ' ' . $this->getDescription($rule) . ' ';
                     $parameterData['setter'] = $this->getFakeFactoryByName('uuid');
                     break;
                 case 'regex':
-                    $parameterData['description'] .= ' ' . $this->getDescription($rule, [':regex' => $arguments[0]]);
-                    $parameterData['setter'] = fn() => $this->getFaker()->regexify($arguments[0]);;
+                    $parameterData['description'] .= ' ' . $this->getDescription($rule, [':regex' => $ruleArguments[0]]);
+                    $parameterData['setter'] = fn() => $this->getFaker()->regexify($ruleArguments[0]);;
                     break;
 
                 /**
                  * Special number types.
                  */
                 case 'digits':
-                    $parameterData['description'] .= ' ' . $this->getDescription($rule, [':digits' => $arguments[0]]);
-                    $parameterData['setter'] = fn() => $this->getFaker()->numerify(str_repeat("#", $arguments[0]));
+                    $parameterData['description'] .= ' ' . $this->getDescription($rule, [':digits' => $ruleArguments[0]]);
+                    $parameterData['setter'] = fn() => $this->getFaker()->numerify(str_repeat("#", $ruleArguments[0]));
                     $parameterData['type'] = 'string';
                     break;
                 case 'digits_between':
-                    $parameterData['description'] .= ' ' . $this->getDescription($rule, [':min' => $arguments[0], ':max' => $arguments[1]]);
-                    $parameterData['setter'] = fn() => $this->getFaker()->numerify(str_repeat("#", rand($arguments[0], $arguments[1])));
+                    $parameterData['description'] .= ' ' . $this->getDescription($rule, [':min' => $ruleArguments[0], ':max' => $ruleArguments[1]]);
+                    $parameterData['setter'] = fn() => $this->getFaker()->numerify(str_repeat("#", rand($ruleArguments[0], $ruleArguments[1])));
                     $parameterData['type'] = 'string';
                     break;
 
@@ -423,29 +450,29 @@ trait ParsesValidationRules
                  */
                 case 'size':
                     $parameterData['description'] .= ' ' . $this->getDescription(
-                            $rule, [':size' => $arguments[0]], $this->getLaravelValidationBaseTypeMapping($parameterData['type'])
+                            $rule, [':size' => $ruleArguments[0]], $this->getLaravelValidationBaseTypeMapping($parameterData['type'])
                         );
-                    $parameterData['setter'] = $this->getDummyValueGenerator($parameterData['type'], ['size' => $arguments[0]]);
+                    $parameterData['setter'] = $this->getDummyValueGenerator($parameterData['type'], ['size' => $ruleArguments[0]]);
                     break;
                 case 'min':
                     $parameterData['description'] .= ' ' . $this->getDescription(
-                            $rule, [':min' => $arguments[0]], $this->getLaravelValidationBaseTypeMapping($parameterData['type'])
+                            $rule, [':min' => $ruleArguments[0]], $this->getLaravelValidationBaseTypeMapping($parameterData['type'])
                         );
-                    $parameterData['setter'] = $this->getDummyDataGeneratorBetween($parameterData['type'], floatval($arguments[0]), fieldName: $parameterData['name']);
+                    $parameterData['setter'] = $this->getDummyDataGeneratorBetween($parameterData['type'], floatval($ruleArguments[0]), fieldName: $parameterData['name']);
                     break;
                 case 'max':
                     $parameterData['description'] .= ' ' . $this->getDescription(
-                            $rule, [':max' => $arguments[0]], $this->getLaravelValidationBaseTypeMapping($parameterData['type'])
+                            $rule, [':max' => $ruleArguments[0]], $this->getLaravelValidationBaseTypeMapping($parameterData['type'])
                         );
-                    $max = min($arguments[0], 25);
+                    $max = min($ruleArguments[0], 25);
                     $parameterData['setter'] = $this->getDummyDataGeneratorBetween($parameterData['type'], 1, $max, $parameterData['name']);
                     break;
                 case 'between':
                     $parameterData['description'] .= ' ' . $this->getDescription(
-                            $rule, [':min' => $arguments[0], ':max' => $arguments[1]], $this->getLaravelValidationBaseTypeMapping($parameterData['type'])
+                            $rule, [':min' => $ruleArguments[0], ':max' => $ruleArguments[1]], $this->getLaravelValidationBaseTypeMapping($parameterData['type'])
                         );
                     // Avoid exponentially complex operations by using the minimum length
-                    $parameterData['setter'] = $this->getDummyDataGeneratorBetween($parameterData['type'], floatval($arguments[0]), floatval($arguments[0]) + 1, $parameterData['name']);
+                    $parameterData['setter'] = $this->getDummyDataGeneratorBetween($parameterData['type'], floatval($ruleArguments[0]), floatval($ruleArguments[0]) + 1, $parameterData['name']);
                     break;
 
                 /**
@@ -464,10 +491,9 @@ trait ParsesValidationRules
                  * Other rules.
                  */
                 case 'in':
-                    // Not using the rule description here because it only says "The attribute is invalid"
-                    $parameterData['description'] .= ' Must be one of ' . w::getListOfValuesAsFriendlyHtmlString($arguments) . ' ';
-                    $parameterData['setter'] = function () use ($arguments) {
-                        return Arr::random($arguments);
+                    $parameterData['enumValues'] = $ruleArguments;
+                    $parameterData['setter'] = function () use ($ruleArguments) {
+                        return Arr::random($ruleArguments);
                     };
                     break;
 
@@ -475,56 +501,56 @@ trait ParsesValidationRules
                  * These rules only add a description. Generating valid examples is too complex.
                  */
                 case 'not_in':
-                    $parameterData['description'] .= ' Must not be one of ' . w::getListOfValuesAsFriendlyHtmlString($arguments) . ' ';
+                    $parameterData['description'] .= ' Must not be one of ' . w::getListOfValuesAsFriendlyHtmlString($ruleArguments) . ' ';
                     break;
                 case 'required_if':
                     $parameterData['description'] .= sprintf(
-                        " This field is required when <code>{$arguments[0]}</code> is %s. ",
-                        w::getListOfValuesAsFriendlyHtmlString(array_slice($arguments, 1))
+                        " This field is required when <code>{$ruleArguments[0]}</code> is %s. ",
+                        w::getListOfValuesAsFriendlyHtmlString(array_slice($ruleArguments, 1))
                     );
                     break;
                 case 'required_unless':
                     $parameterData['description'] .= sprintf(
-                        " This field is required unless <code>{$arguments[0]}</code> is in %s. ",
-                        w::getListOfValuesAsFriendlyHtmlString(array_slice($arguments, 1))
+                        " This field is required unless <code>{$ruleArguments[0]}</code> is in %s. ",
+                        w::getListOfValuesAsFriendlyHtmlString(array_slice($ruleArguments, 1))
                     );
                     break;
                 case 'required_with':
                     $parameterData['description'] .= sprintf(
                         " This field is required when %s is present. ",
-                        w::getListOfValuesAsFriendlyHtmlString($arguments)
+                        w::getListOfValuesAsFriendlyHtmlString($ruleArguments)
                     );
                     break;
                 case 'required_without':
                     $parameterData['description'] .= sprintf(
                         " This field is required when %s is not present. ",
-                        w::getListOfValuesAsFriendlyHtmlString($arguments)
+                        w::getListOfValuesAsFriendlyHtmlString($ruleArguments)
                     );
                     break;
                 case 'required_with_all':
                     $parameterData['description'] .= sprintf(
                         " This field is required when %s are present. ",
-                        w::getListOfValuesAsFriendlyHtmlString($arguments, "and")
+                        w::getListOfValuesAsFriendlyHtmlString($ruleArguments, "and")
                     );
                     break;
                 case 'required_without_all':
                     $parameterData['description'] .= sprintf(
-                    " This field is required when none of %s are present. ",
-                    w::getListOfValuesAsFriendlyHtmlString($arguments, "and")
+                        " This field is required when none of %s are present. ",
+                        w::getListOfValuesAsFriendlyHtmlString($ruleArguments, "and")
                     );
                     break;
-                case 'accepted_if':
-                    $parameterData['type'] = 'boolean';
-                    $parameterData['description'] .= " Must be accepted when <code>$arguments[0]</code> is " . w::getListOfValuesAsFriendlyHtmlString(array_slice($arguments, 1));
-                    $parameterData['setter'] = fn() => true;
-                    break;
                 case 'same':
-                    $parameterData['description'] .= " The value and <code>{$arguments[0]}</code> must match.";
+                    $parameterData['description'] .= " The value and <code>{$ruleArguments[0]}</code> must match.";
                     break;
                 case 'different':
-                    $parameterData['description'] .= " The value and <code>{$arguments[0]}</code> must be different.";
+                    $parameterData['description'] .= " The value and <code>{$ruleArguments[0]}</code> must be different.";
                     break;
-
+                case 'nullable':
+                    $parameterData['nullable'] = true;
+                    break;
+                case 'exists':
+                    $parameterData['description'] .= " The <code>{$ruleArguments[1]}</code> of an existing record in the {$ruleArguments[0]} table.";
+                    break;
                 default:
                     // Other rules not supported
                     break;
@@ -533,11 +559,13 @@ trait ParsesValidationRules
             throw CouldntProcessValidationRule::forParam($parameterData['name'], $rule, $e);
         }
 
+        $parameterData['description'] = trim($parameterData['description']);
         return true;
     }
 
     /**
      * Parse a string rule into the base rule and arguments.
+     * eg "in:1,2" becomes ["in", ["1", "2"]]
      * Laravel validation rules are specified in the format {rule}:{arguments}
      * Arguments are separated by commas.
      * For instance the rule "max:3" states that the value may only be three letters.
@@ -550,7 +578,7 @@ trait ParsesValidationRules
     {
         $ruleArguments = [];
 
-        if (strpos($rule, ':') !== false) {
+        if (str_contains($rule, ':')) {
             [$rule, $argumentsString] = explode(':', $rule, 2);
 
             // These rules can have commas in their arguments, so we don't split on commas
@@ -577,7 +605,7 @@ trait ParsesValidationRules
                     : null;
             }
         } else if (!is_null($parameterData['example']) && $parameterData['example'] !== self::$MISSING_VALUE) {
-            if($parameterData['example'] === 'No-example' && !$parameterData['required']){
+            if ($parameterData['example'] === 'No-example' && !$parameterData['required']) {
                 return null;
             }
             // Casting again is important since values may have been cast to string in the validator
@@ -629,8 +657,11 @@ trait ParsesValidationRules
                 // 2. If `users.<name>` exists, `users` is an `object`
                 // 3. Otherwise, default to `object`
                 // Important: We're iterating in reverse, to ensure we set child items before parent items
-                // (assuming the user specified parents first, which is the more common thing)
-                if ($childKey = Arr::first($allKeys, fn($key) => Str::startsWith($key, "$name.*"))) {
+                // (assuming the user specified parents first, which is the more common thing)y
+                if (Arr::first($allKeys, fn($key) => Str::startsWith($key, "$name.*."))) {
+                    $details['type'] = 'object[]';
+                    unset($details['setter']);
+                } else if ($childKey = Arr::first($allKeys, fn($key) => Str::startsWith($key, "$name.*"))) {
                     $childType = ($converted[$childKey] ?? $parameters[$childKey])['type'];
                     $details['type'] = "{$childType}[]";
                 } else { // `array` types default to `object` if no subtype is specified
@@ -656,14 +687,14 @@ trait ParsesValidationRules
         foreach ($parameters as $name => $details) {
             if (Str::endsWith($name, '.*')) {
                 // The user might have set the example via bodyParameters()
-                $hasExample = $this->examplePresent($details);
+                $exampleWasSpecified = $this->examplePresent($details);
 
                 // Change cars.*.dogs.things.*.* with type X to cars.*.dogs.things with type X[][]
                 while (Str::endsWith($name, '.*')) {
                     $details['type'] .= '[]';
                     $name = substr($name, 0, -2);
 
-                    if ($hasExample) {
+                    if ($exampleWasSpecified) {
                         $details['example'] = [$details['example']];
                     } else if (isset($details['setter'])) {
                         $previousSetter = $details['setter'];
@@ -771,11 +802,21 @@ trait ParsesValidationRules
             return "Must match the regex {$arguments[':regex']}.";
         }
 
-        $description = trans("validation.{$rule}");
-        // For rules that can apply to multiple types (eg 'max' rule), Laravel returns an array of possible messages
+        $translationString = "validation.{$rule}";
+        $description = trans($translationString);
+
+        // For rules that can apply to multiple types (eg 'max' rule), There is an array of possible messages
         // 'numeric' => 'The :attribute must not be greater than :max'
         // 'file' => 'The :attribute must have a size less than :max kilobytes'
-        if (is_array($description)) {
+        // Depending on the translation engine, trans may return the array, or it will fail to translate the string
+        // and will need to be called with the baseType appended.
+        if ($description === $translationString) {
+            $translationString = "{$translationString}.{$baseType}";
+            $translated = trans($translationString);
+            if ($translated !== $translationString) {
+                $description = $translated;
+            }
+        } elseif (is_array($description)) {
             $description = $description[$baseType];
         }
 
@@ -824,5 +865,17 @@ trait ParsesValidationRules
     protected function shouldCastUserExample()
     {
         return false;
+    }
+
+    protected function warnAboutMissingCustomParameterData(string $parameter, array $customParameterData): array
+    {
+        $parameterPlusDot = $parameter . '.';
+        if (count($customParameterData) && !isset($customParameterData[$parameter])
+            && !Arr::first(array_keys($customParameterData), fn($key) => str_starts_with($key, $parameterPlusDot))
+        ) {
+            c::debug($this->getMissingCustomDataMessage($parameter));
+        }
+
+        return $customParameterData;
     }
 }
