@@ -2,15 +2,15 @@
 
 namespace Knuckles\Scribe\Tools;
 
+use Illuminate\Support\Str;
 use Knuckles\Scribe\Tools\ConsoleOutputUtils as c;
 
 class DocumentationConfig
 {
-    private $data;
+    public array $data;
 
     public function __construct(array $config = [])
     {
-        $config['router'] = $this->getRouter($config);
         $this->data = $config;
     }
 
@@ -28,21 +28,18 @@ class DocumentationConfig
         return data_get($this->data, $key, $default);
     }
 
-    private function getRouter(array $config): string
+    public function outputIsStatic(): bool
     {
-        if ($router = data_get($config, 'router', null)) {
-            if (!in_array($router, ['dingo', 'laravel'])) {
-                throw new \InvalidArgumentException("Unknown `router` config value: $router");
-            }
-            return $router;
-        }
+        return !$this->outputRoutedThroughLaravel();
+    }
 
-        if (class_exists(\Dingo\Api\Routing\Router::class)) {
-            c::info('Detected Dingo API router');
-            return 'dingo';
-        }
+    public function outputRoutedThroughLaravel(): bool
+    {
+        return Str::is(['laravel', 'external_laravel'], $this->get('type'));
+    }
 
-        return 'laravel';
-
+    public function outputIsExternal(): bool
+    {
+        return Str::is(['external_static', 'external_laravel'], $this->get('type'));
     }
 }
