@@ -8,15 +8,30 @@ use Attribute;
 class Subgroup
 {
     public function __construct(
-        public string $name,
+        public mixed $name,
         public ?string $description = '',
     ){
+    }
+
+    protected function getName(): string
+    {
+        if (is_string($this->name)) {
+            return $this->name;
+        }
+
+        if (interface_exists('BackedEnum') && is_a($this->name, 'BackedEnum')) {
+            return $this->name->value;
+        }
+
+        throw new \InvalidArgumentException(
+            'The name property of a subgroup must be either a PHP Backed Enum or a string'
+        );
     }
 
     public function toArray()
     {
         return [
-            "subgroup" => $this->name,
+            "subgroup" => $this->getName(),
             "subgroupDescription" => $this->description,
         ];
     }
