@@ -21,7 +21,7 @@ trait InstantiatesExampleModels
      */
     protected function instantiateExampleModel(
         ?string $type = null, array $factoryStates = [],
-        array   $relations = [], ?ReflectionFunctionAbstract $transformationMethod = null
+        array   $relations = [], ?ReflectionFunctionAbstract $transformationMethod = null, array $withCount = [],
     )
     {
         // If the API Resource uses an empty resource, there won't be an example model
@@ -30,9 +30,11 @@ trait InstantiatesExampleModels
 
         if ($type == null) {
             $parameter = Arr::first($transformationMethod->getParameters());
-            if ($parameter->hasType() && !$parameter->getType()->isBuiltin() && class_exists($parameter->getType()->getName())) {
+            $parameterType = $parameter->hasType() ? $parameter->getType() : null;
+            if ($parameterType instanceof \ReflectionNamedType &&
+                !$parameterType->isBuiltin() && class_exists($parameterType->getName())) {
                 // Ladies and gentlemen, we have a type!
-                $type = $parameter->getType()->getName();
+                $type = $parameterType->getName();
             }
         }
         if ($type == null) {
@@ -42,7 +44,7 @@ trait InstantiatesExampleModels
         $configuredStrategies = $this->config->get('examples.models_source', ['factoryCreate', 'factoryMake', 'databaseFirst']);
 
         $strategies = [
-            'factoryCreate' => fn() => $this->getExampleModelFromFactoryCreate($type, $factoryStates, $relations),
+            'factoryCreate' => fn() => $this->getExampleModelFromFactoryCreate($type, $factoryStates, $relations, $withCount),
             'factoryMake' => fn() => $this->getExampleModelFromFactoryMake($type, $factoryStates, $relations),
             'databaseFirst' => fn() => $this->getExampleModelFromDatabaseFirst($type, $relations),
         ];
@@ -53,7 +55,7 @@ trait InstantiatesExampleModels
                 if ($model) return $model;
             } catch (Throwable $e) {
                 c::warn("Couldn't get example model for {$type} via $strategyName.");
-                e::dumpExceptionIfVerbose($e, true);
+                e::dumpExceptionIfVerbose($e);
             }
         }
 
@@ -63,13 +65,19 @@ trait InstantiatesExampleModels
     /**
      * @param class-string $type
      * @param string[] $factoryStates
+     * @param string[] $relations
+     * @param string[] $withCount
      *
      * @return \Illuminate\Database\Eloquent\Model|null
      */
-    protected function getExampleModelFromFactoryCreate(string $type, array $factoryStates = [], array $relations = [])
+    protected function getExampleModelFromFactoryCreate(string $type, array $factoryStates = [], array $relations = [], array $withCount = [])
     {
-        $factory = Utils::getModelFactory($type, $factoryStates, $relations);
-        return $factory->create()->load($relations);
+        // Since $relations and $withCount refer to the same underlying relationships in the model,
+        // combining them ensures that all required relationships are initialized when passed to the factory.
+        $allRelations = array_unique(array_merge($relations, $withCount));
+
+        $factory = Utils::getModelFactory($type, $factoryStates, $allRelations);
+        return $factory->create()->refresh()->load($relations)->loadCount($withCount);
     }
 
     /**
