@@ -3,10 +3,12 @@
 namespace Knuckles\Scribe\Extracting\Shared\ValidationRulesFinders;
 
 use PhpParser\Node;
+use PhpParser\NodeFinder;
 
 /**
  * This class looks for
  *   $validator = Validator::make($request, ...)
+ *   Validator::make($request, ...)->validate()
  *
  * The variable names (`$validator` and `$request`) don't matter.
  */
@@ -15,20 +17,19 @@ class ValidatorMake
     public static function find(Node $node)
     {
         // Make sure it's an assignment
-        if (!($node instanceof Node\Stmt\Expression)
-            || !($node->expr instanceof Node\Expr\Assign)) {
+        if (! ($node instanceof Node\Stmt\Expression)) {
             return;
         }
 
-        $expr = $node->expr->expr; // Get the expression on the RHS
+        $validatorNode = (new NodeFinder)->findFirst($node, function ($node): bool {
+            return $node instanceof Node\Expr\StaticCall
+                && ! empty($node->class->name)
+                && str_ends_with($node->class->name, 'Validator')
+                && $node->name->name == 'make';
+        });
 
-        if (
-            $expr instanceof Node\Expr\StaticCall
-            && !empty($expr->class->parts)
-            && end($expr->class->parts) == "Validator"
-            && $expr->name->name == "make"
-        ) {
-            return $expr->args[1]->value;
+        if ($validatorNode instanceof Node\Expr\StaticCall) {
+            return $validatorNode->args[1]->value;
         }
     }
 }
