@@ -11,12 +11,11 @@ class GetFromDocBlocks extends Strategy
 {
     public function __invoke(ExtractedEndpointData $endpointData, array $routeRules = []): array
     {
-        $group = substr($endpointData->controller->name, 0,20);
         $docBlocks = RouteDocBlocker::getDocBlocksFromRoute($endpointData->route);
         $methodDocBlock = $docBlocks['method'];
         $classDocBlock = $docBlocks['class'];
 
-        return array_merge($this->getMetadataFromDocBlock($methodDocBlock, $classDocBlock), ['groupName'=> $group]);
+        return $this->getMetadataFromDocBlock($methodDocBlock, $classDocBlock);
     }
 
     public function getMetadataFromDocBlock(DocBlock $methodDocBlock, DocBlock $classDocBlock): array
@@ -30,6 +29,7 @@ class GetFromDocBlocks extends Strategy
             'subgroupDescription' => $this->getEndpointSubGroupDescription($methodDocBlock, $classDocBlock),
             'title' => $title ?: $methodDocBlock->getShortDescription(),
             'description' => $methodDocBlock->getLongDescription()->getContents(),
+            'deprecated' => $this->getDeprecatedStatusFromDocBlock($methodDocBlock, $classDocBlock),
         ];
         if (!is_null($authStatus = $this->getAuthStatusFromDocBlock($methodDocBlock, $classDocBlock))) {
             $metadata['authenticated'] = $authStatus;
@@ -37,7 +37,7 @@ class GetFromDocBlocks extends Strategy
         return $metadata;
     }
 
-    protected function getAuthStatusFromDocBlock(DocBlock $methodDocBlock, DocBlock $classDocBlock = null): ?bool
+    protected function getAuthStatusFromDocBlock(DocBlock $methodDocBlock, ?DocBlock $classDocBlock = null): ?bool
     {
         foreach ($methodDocBlock->getTags() as $tag) {
             if (strtolower($tag->getName()) === 'authenticated') {
@@ -52,6 +52,17 @@ class GetFromDocBlocks extends Strategy
         return $classDocBlock
             ? $this->getAuthStatusFromDocBlock($classDocBlock)
             : null;
+    }
+
+    protected function getDeprecatedStatusFromDocBlock(DocBlock $methodDocBlock, ?DocBlock $classDocBlock = null): bool
+    {
+        foreach ($methodDocBlock->getTags() as $tag) {
+            if (strtolower($tag->getName()) === 'deprecated') {
+                return true;
+            }
+        }
+
+        return $classDocBlock && $this->getDeprecatedStatusFromDocBlock($classDocBlock);
     }
 
     /**
