@@ -2,17 +2,14 @@
 
 namespace Knuckles\Scribe\Extracting\Strategies\Responses;
 
-use Illuminate\Support\Facades\Config;
-use Knuckles\Camel\Extraction\ExtractedEndpointData;
-use Dingo\Api\Dispatcher;
-use Dingo\Api\Routing\Route as DingoRoute;
 use Exception;
 use Illuminate\Contracts\Http\Kernel;
 use Illuminate\Http\Request;
-use Illuminate\Http\Response;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Routing\Route;
+use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Str;
+use Knuckles\Camel\Extraction\ExtractedEndpointData;
 use Knuckles\Scribe\Extracting\DatabaseTransactionHelpers;
 use Knuckles\Scribe\Extracting\ParamHelpers;
 use Knuckles\Scribe\Extracting\Strategies\Strategy;
@@ -20,6 +17,8 @@ use Knuckles\Scribe\Tools\ConsoleOutputUtils as c;
 use Knuckles\Scribe\Tools\ErrorHandlingUtils as e;
 use Knuckles\Scribe\Tools\Globals;
 use Knuckles\Scribe\Tools\Utils;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\StreamedResponse;
 
 /**
  * Make a call to the route and retrieve its response.
@@ -30,28 +29,23 @@ class ResponseCalls extends Strategy
 
     protected array $previousConfigs = [];
 
-    public function __invoke(ExtractedEndpointData $endpointData, array $routeRules = []): ?array
+    public function __invoke(ExtractedEndpointData $endpointData, array $settings = []): ?array
     {
-        return $this->makeResponseCallIfConditionsPass($endpointData, $routeRules);
-    }
-
-    public function makeResponseCallIfConditionsPass(ExtractedEndpointData $endpointData, array $routeRules): ?array
-    {
-        $rulesToApply = $routeRules['response_calls'] ?? [];
-        if (!$this->shouldMakeApiCall($endpointData, $rulesToApply)) {
+        // Don't attempt a response call if there are already successful responses
+        if ($endpointData->responses->hasSuccessResponse()) {
             return null;
         }
 
-        return $this->makeResponseCall($endpointData, $rulesToApply);
+        return $this->makeResponseCall($endpointData, $settings);
     }
 
-    public function makeResponseCall(ExtractedEndpointData $endpointData, array $rulesToApply): ?array
+    public function makeResponseCall(ExtractedEndpointData $endpointData, array $settings): ?array
     {
-        $this->configureEnvironment($rulesToApply);
+        $this->configureEnvironment($settings);
 
         // Mix in parsed parameters with manually specified parameters.
-        $bodyParameters = array_merge($endpointData->cleanBodyParameters, $rulesToApply['bodyParams'] ?? []);
-        $queryParameters = array_merge($endpointData->cleanQueryParameters, $rulesToApply['queryParams'] ?? []);
+        $bodyParameters = array_merge($endpointData->cleanBodyParameters, $settings['bodyParams'] ?? []);
+        $queryParameters = array_merge($endpointData->cleanQueryParameters, $settings['queryParams'] ?? []);
         $urlParameters = $endpointData->cleanUrlParameters;
         $headers = $endpointData->headers;
 
@@ -72,17 +66,17 @@ class ResponseCalls extends Strategy
             }
         }
 
-        $hardcodedFileParams = $rulesToApply['fileParams'] ?? [];
+        $hardcodedFileParams = $settings['fileParams'] ?? [];
         $hardcodedFileParams = collect($hardcodedFileParams)->map(function ($filePath) {
             $fileName = basename($filePath);
             return new UploadedFile(
-                $filePath, $fileName, mime_content_type($filePath), 0, false
+                $filePath, $fileName, mime_content_type($filePath), test: true
             );
         })->toArray();
         $fileParameters = array_merge($endpointData->fileParameters, $hardcodedFileParams);
 
         $request = $this->prepareRequest(
-            $endpointData->route, $endpointData->uri, $rulesToApply, $urlParameters,
+            $endpointData->route, $endpointData->uri, $settings, $urlParameters,
             $bodyParameters, $queryParameters, $fileParameters, $headers
         );
 
@@ -90,10 +84,13 @@ class ResponseCalls extends Strategy
 
         try {
             $response = $this->makeApiCall($request, $endpointData->route);
+
+            $this->runPostRequestHook($request, $endpointData, $response);
+
             $response = [
                 [
                     'status' => $response->getStatusCode(),
-                    'content' => $response->getContent(),
+                    'content' => $this->getContentFromResponse($response),
                     'headers' => $this->getResponseHeaders($response),
                 ],
             ];
@@ -110,19 +107,19 @@ class ResponseCalls extends Strategy
     }
 
     /**
-     * @param array $rulesToApply
+     * @param array $settings
      *
      * @return void
      */
-    private function configureEnvironment(array $rulesToApply)
+    private function configureEnvironment(array $settings)
     {
         $this->startDbTransaction();
-        $this->setLaravelConfigs($rulesToApply['config'] ?? []);
+        $this->setLaravelConfigs($settings['config'] ?? []);
     }
 
     /**
      * @param Route $route
-     * @param array $rulesToApply
+     * @param array $settings
      * @param array $urlParams
      * @param array $bodyParams
      * @param array $queryParams
@@ -132,13 +129,15 @@ class ResponseCalls extends Strategy
      *
      * @return Request
      */
-    protected function prepareRequest(Route $route, string $url, array $rulesToApply, array $urlParams,
-        array $bodyParams, array $queryParams, array $fileParameters, array $headers): Request
+    protected function prepareRequest(
+        Route $route, string $url, array $settings, array $urlParams,
+        array $bodyParams, array $queryParams, array $fileParameters, array $headers
+    ): Request
     {
         $uri = Utils::getUrlWithBoundParameters($url, $urlParams);
         $routeMethods = $this->getMethods($route);
         $method = array_shift($routeMethods);
-        $cookies = $rulesToApply['cookies'] ?? [];
+        $cookies = $settings['cookies'] ?? [];
 
         // Note that we initialise the request with the bodyParams here
         // and later still add them to the ParameterBag (`addBodyParameters`)
@@ -169,6 +168,13 @@ class ResponseCalls extends Strategy
         }
     }
 
+    protected function runPostRequestHook(Request $request, ExtractedEndpointData $endpointData, mixed $response): void
+    {
+        if (is_callable(Globals::$__afterResponseCall)) {
+            call_user_func_array(Globals::$__afterResponseCall, [$request, $endpointData, $response]);
+        }
+    }
+
     private function setLaravelConfigs(array $config)
     {
         if (empty($config)) {
@@ -192,47 +198,6 @@ class ResponseCalls extends Strategy
     {
         $this->endDbTransaction();
         $this->rollbackLaravelConfigChanges();
-    }
-
-    /**
-     * @param Request $request
-     *
-     * @return \Illuminate\Http\JsonResponse|mixed
-     */
-    public function callDingoRoute(Request $request, Route $route)
-    {
-        /** @var \Dingo\Api\Dispatcher $dispatcher */
-        $dispatcher = app(\Dingo\Api\Dispatcher::class);
-
-        /** @var DingoRoute $route */
-        $dispatcher->version($route->versions()[0]);
-        foreach ($request->headers as $header => $value) {
-            $dispatcher->header($header, $value);
-        }
-
-        // set domain and body parameters
-        $dispatcher->on($request->header('SERVER_NAME'))
-            ->with($request->request->all());
-
-        // set URL and query parameters
-        $uri = $request->getRequestUri();
-        $query = $request->getQueryString();
-        if (!empty($query)) {
-            $uri .= "?$query";
-        }
-
-        $response = call_user_func_array(
-            [$dispatcher, strtolower($request->method())],
-            [$uri]
-        );
-
-        // the response from the Dingo dispatcher is the 'raw' response from the controller,
-        // so we have to ensure it's JSON first
-        if (!$response instanceof Response) {
-            $response = response()->json($response);
-        }
-
-        return $response;
     }
 
     public function getMethods(Route $route): array
@@ -276,67 +241,22 @@ class ResponseCalls extends Strategy
      *
      * @param Route $route
      *
-     * @return \Illuminate\Http\JsonResponse|mixed|\Symfony\Component\HttpFoundation\Response
+     * @return Response
      * @throws Exception
      */
     protected function makeApiCall(Request $request, Route $route)
     {
-        if ($this->config->get('router') == 'dingo') {
-            $response = $this->callDingoRoute($request, $route);
-        } else {
-            $response = $this->callLaravelOrLumenRoute($request);
-        }
-
-        return $response;
+        return $this->callLaravelRoute($request);
     }
 
-    protected function callLaravelOrLumenRoute(Request $request): \Symfony\Component\HttpFoundation\Response
+    protected function callLaravelRoute(Request $request): Response
     {
-        // Confirm we're running in Laravel, not Lumen
-        if (app()->bound(Kernel::class)) {
-            /** @var \Illuminate\Foundation\Http\Kernel $kernel */
-            $kernel = app(Kernel::class);
-            $response = $kernel->handle($request);
-            $kernel->terminate($request, $response);
-        } else {
-            // Handle the request using the Lumen application.
-            /** @var \Laravel\Lumen\Application $app */
-            $app = app();
-            $app->bind('request', function () use ($request) {
-                return $request;
-            });
-            $response = $app->handle($request);
-        }
+        /** @var \Illuminate\Foundation\Http\Kernel $kernel */
+        $kernel = app(Kernel::class);
+        $response = $kernel->handle($request);
+        $kernel->terminate($request, $response);
 
         return $response;
-    }
-
-    protected function shouldMakeApiCall(ExtractedEndpointData $endpointData, array $rulesToApply): bool
-    {
-        $allowedMethods = $rulesToApply['methods'] ?? [];
-        if (empty($allowedMethods)) {
-            return false;
-        }
-
-        // Don't attempt a response call if there are already successful responses
-        if ($endpointData->responses->hasSuccessResponse()) {
-            return false;
-        }
-
-        if (is_string($allowedMethods) && $allowedMethods == '*') {
-            return true;
-        }
-
-        if (array_search('*', $allowedMethods) !== false) {
-            return true;
-        }
-
-        $routeMethods = $this->getMethods($endpointData->route);
-        if (in_array(array_shift($routeMethods), $allowedMethods)) {
-            return true;
-        }
-
-        return false;
     }
 
     /**
@@ -367,5 +287,63 @@ class ResponseCalls extends Strategy
         }
 
         return $formattedHeaders;
+    }
+
+    /**
+     * @param array $only The routes which this strategy should be applied to. Can not be specified with $except.
+     *   Specify route names ("users.index", "users.*"), or method and path ("GET *", "POST /safe/*").
+     * @param array $except The routes which this strategy should be applied to. Can not be specified with $only.
+     *   Specify route names ("users.index", "users.*"), or method and path ("GET *", "POST /safe/*").
+     * @param array $config Any extra Laravel config() values to before starting the response call.
+     * @param array $queryParams Query params to always send with the response call. Key-value array.
+     * @param array $bodyParams Body params to always send with the response call. Key-value array.
+     * @param array $fileParams File params to always send with the response call. Key-value array. Key is param name, value is file path.
+     * @param array $cookies Cookies to always send with the response call. Key-value array.
+     * @return array
+     */
+    public static function withSettings(
+        array $only = [],
+        array $except = [],
+        array $config = [],
+        array $queryParams = [],
+        array $bodyParams = [],
+        array $fileParams = [
+            // 'key' => 'storage/app/image.png',
+        ],
+        array $cookies = [],
+    ): array
+    {
+        return static::wrapWithSettings(
+            only: $only,
+            except: $except,
+            otherSettings: compact(
+                'config',
+                'queryParams',
+                'bodyParams',
+                'fileParams',
+                'cookies',
+            ));
+    }
+
+    protected function getContentFromResponse(Response $response): string|false
+    {
+        if (!$response instanceof StreamedResponse) {
+            return $response->getContent();
+        }
+
+        // For streamed responses, the content is null, and only output directly via "echo" when we call "sendContent".
+        // We use output buffering to capture the output into a new fake response.
+        $renderedResponse = new Response('', $response->getStatusCode());
+        $originalCallback = $response->getCallback();
+        $response->setCallback(function () use ($originalCallback, $renderedResponse) {
+            ob_start(function ($output) use ($renderedResponse) {
+                $renderedResponse->setContent($output);
+            });
+            $originalCallback();
+            ob_end_flush();
+        });
+        $response->sendContent();
+        $renderedResponse->headers = $response->headers;
+        return $renderedResponse->getContent();
     }
 }
