@@ -2,23 +2,21 @@
 
 namespace Knuckles\Scribe\Matching;
 
-use Dingo\Api\Routing\RouteCollection;
 use Illuminate\Routing\Route;
 use Illuminate\Support\Facades\Route as RouteFacade;
 use Illuminate\Support\Str;
+use Knuckles\Scribe\Tools\RoutePatternMatcher;
 
 class RouteMatcher implements RouteMatcherInterface
 {
-    public function getRoutes(array $routeRules = [], string $router = 'laravel'): array
+    public function getRoutes(array $routeRules = []): array
     {
-        $usingDingoRouter = strtolower($router) == 'dingo';
-
-        return $this->getRoutesToBeDocumented($routeRules, $usingDingoRouter);
+        return $this->getRoutesToBeDocumented($routeRules);
     }
 
-    private function getRoutesToBeDocumented(array $routeRules, bool $usingDingoRouter = false): array
+    private function getRoutesToBeDocumented(array $routeRules): array
     {
-        $allRoutes = $this->getAllRoutes($usingDingoRouter);
+        $allRoutes = $this->getAllRoutes();
 
         $matchedRoutes = [];
 
@@ -26,15 +24,11 @@ class RouteMatcher implements RouteMatcherInterface
             $includes = $routeRule['include'] ?? [];
 
             foreach ($allRoutes as $route) {
-                if (is_array($route)) {
-                    $route = new LumenRouteAdapter($route);
-                }
-
                 if ($this->shouldExcludeRoute($route, $routeRule)) {
                     continue;
                 }
 
-                if ($this->shouldIncludeRoute($route, $routeRule, $includes, $usingDingoRouter)) {
+                if ($this->shouldIncludeRoute($route, $routeRule, $includes)) {
                     $matchedRoutes[] = new MatchedRoute($route, $routeRule['apply'] ?? []);
                 }
             }
@@ -43,38 +37,21 @@ class RouteMatcher implements RouteMatcherInterface
         return $matchedRoutes;
     }
 
-    private function getAllRoutes(bool $usingDingoRouter)
+    private function getAllRoutes()
     {
-        if (! $usingDingoRouter) {
-            return RouteFacade::getRoutes();
-        }
-
-        /** @var \Dingo\Api\Routing\Router $router */
-        $router = app(\Dingo\Api\Routing\Router::class);
-        $allRouteCollections = $router->getRoutes();
-
-        return collect($allRouteCollections)
-            ->flatMap(function (RouteCollection $collection) {
-                return $collection->getRoutes();
-            })->toArray();
+        return RouteFacade::getRoutes();
     }
 
-    private function shouldIncludeRoute(Route $route, array $routeRule, array $mustIncludes, bool $usingDingoRouter): bool
+    private function shouldIncludeRoute(Route $route, array $routeRule, array $mustIncludes): bool
     {
-        if (Str::is($mustIncludes, $route->getName()) || Str::is($mustIncludes, $route->uri())) {
+        if (RoutePatternMatcher::matches($route, $mustIncludes)) {
             return true;
-        }
-
-        $matchesVersion = true;
-        if ($usingDingoRouter) {
-            $matchesVersion = !empty(array_intersect($route->versions(), $routeRule['match']['versions'] ?? []));
         }
 
         $domainsToMatch = $routeRule['match']['domains'] ?? [];
         $pathsToMatch = $routeRule['match']['prefixes'] ?? [];
 
-        return Str::is($domainsToMatch, $route->getDomain()) && Str::is($pathsToMatch, $route->uri())
-            && $matchesVersion;
+        return Str::is($domainsToMatch, $route->getDomain()) && Str::is($pathsToMatch, $route->uri());
     }
 
     private function shouldExcludeRoute(Route $route, array $routeRule): bool
@@ -90,7 +67,6 @@ class RouteMatcher implements RouteMatcherInterface
             $excludes[] = 'telescope/*';
         }
 
-        return Str::is($excludes, $route->getName())
-            || Str::is($excludes, $route->uri());
+        return RoutePatternMatcher::matches($route, $excludes);
     }
 }
