@@ -3,30 +3,22 @@
 namespace Knuckles\Scribe\Writing;
 
 use Illuminate\Support\Facades\Storage;
+use Knuckles\Camel\Output\OutputEndpointData;
 use Knuckles\Scribe\Tools\ConsoleOutputUtils as c;
 use Knuckles\Scribe\Tools\DocumentationConfig;
 use Knuckles\Scribe\Tools\Globals;
+use Knuckles\Scribe\Tools\PathConfig;
 use Knuckles\Scribe\Tools\Utils;
-use MetaFox\Platform\ApiDoc\MdxWriter;
 use Symfony\Component\Yaml\Yaml;
 
 class Writer
 {
-    /**
-     * The "name" of this docs instance. By default, it is "scribe".
-     * Used for multi-docs.
-     */
-    public string $docsName;
+    protected bool $isStatic;
+    protected bool $isExternal;
 
-    private DocumentationConfig $config;
+    protected ?string $staticTypeOutputPath;
 
-    private bool $isStatic;
-
-    private string $markdownOutputPath;
-
-    private ?string $staticTypeOutputPath;
-
-    private ?string $laravelTypeOutputPath;
+    protected ?string $laravelTypeOutputPath;
     protected array $generatedFiles = [
         'postman' => null,
         'openapi' => null,
@@ -39,42 +31,40 @@ class Writer
         ],
     ];
 
-    private string $laravelAssetsPath;
+    protected string $laravelAssetsPath;
 
-    public function __construct(DocumentationConfig $config = null, $docsName = 'scribe')
+    public function __construct(protected DocumentationConfig $config, public PathConfig $paths)
     {
-        $this->docsName = $docsName;
+        $this->isStatic = $this->config->outputIsStatic();
+        $this->isExternal = $this->config->outputIsExternal();
 
-        // If no config is injected, pull from global, for easier testing.
-        $this->config = $config ?: new DocumentationConfig(config($docsName));
-
-        $this->isStatic = $this->config->get('type') === 'static';
-        $this->markdownOutputPath = ".{$docsName}"; //.scribe by default
         $this->laravelTypeOutputPath = $this->getLaravelTypeOutputPath();
         $this->staticTypeOutputPath = rtrim($this->config->get('static.output_path', 'public/docs'), '/');
 
         $this->laravelAssetsPath = $this->config->get('laravel.assets_directory')
             ? '/' . $this->config->get('laravel.assets_directory')
-            : "/vendor/$this->docsName";
+            : "/vendor/" . $this->paths->outputPath();
     }
 
     /**
-     * @param array[] $groupedEndpoints
+     * @param array<string,array> $groupedEndpoints
      */
-    public function writeDocs(array $groupedEndpoints)
+    public function writeDocs(array $groupedEndpoints): void
     {
         // The static assets (js/, css/, and images/) always go in public/docs/.
         // For 'static' docs, the output files (index.html, collection.json) go in public/docs/.
         // For 'laravel' docs, the output files (index.blade.php, collection.json)
         // go in resources/views/scribe/ and storage/app/scribe/ respectively.
 
-//        app()->makeWith(MdxWriter::class, ['config'=> $this->config])->generateMdx($groupedEndpoints);
-//
-        $this->writeHtmlDocs($groupedEndpoints);
-
-        $this->writePostmanCollection($groupedEndpoints);
-
-        $this->writeOpenAPISpec($groupedEndpoints);
+        if ($this->isExternal) {
+            $this->writeOpenAPISpec($groupedEndpoints);
+            $this->writePostmanCollection($groupedEndpoints);
+            $this->writeExternalHtmlDocs();
+        } else {
+            $this->writeHtmlDocs($groupedEndpoints);
+            $this->writePostmanCollection($groupedEndpoints);
+            $this->writeOpenAPISpec($groupedEndpoints);
+        }
 
         $this->runAfterGeneratingHook();
     }
@@ -89,8 +79,9 @@ class Writer
                 $collectionPath = "{$this->staticTypeOutputPath}/collection.json";
                 file_put_contents($collectionPath, $collection);
             } else {
-                Storage::disk('local')->put("{$this->docsName}/collection.json", $collection);
-                $collectionPath = Storage::disk('local')->path("$this->docsName/collection.json");
+                $outputPath = $this->paths->outputPath('collection.json');
+                Storage::disk('local')->put($outputPath, $collection);
+                $collectionPath = Storage::disk('local')->path($outputPath);
             }
 
             c::success("Wrote Postman collection to: {$this->makePathFriendly($collectionPath)}");
@@ -100,16 +91,18 @@ class Writer
 
     protected function writeOpenAPISpec(array $parsedRoutes): void
     {
-        if ($this->config->get('openapi.enabled', false)) {
+        if ($this->config->get('openapi.enabled', false) || $this->isExternal) {
             c::info('Generating OpenAPI specification');
 
             $spec = $this->generateOpenAPISpec($parsedRoutes);
             if ($this->isStatic) {
+                Utils::makeDirectoryRecursive($this->staticTypeOutputPath);
                 $specPath = "{$this->staticTypeOutputPath}/openapi.yaml";
                 file_put_contents($specPath, $spec);
             } else {
-                Storage::disk('local')->put("{$this->docsName}/openapi.yaml", $spec);
-                $specPath = Storage::disk('local')->path("$this->docsName/openapi.yaml");
+                $outputPath = $this->paths->outputPath('openapi.yaml');
+                Storage::disk('local')->put($outputPath, $spec);
+                $specPath = Storage::disk('local')->path($outputPath);
             }
 
             c::success("Wrote OpenAPI specification to: {$this->makePathFriendly($specPath)}");
@@ -136,7 +129,7 @@ class Writer
                 data_set($collection, $key, $value);
             }
         }
-        return json_encode($collection, JSON_PRETTY_PRINT);
+        return json_encode($collection, JSON_PRETTY_PRINT|JSON_INVALID_UTF8_SUBSTITUTE);
     }
 
     /**
@@ -148,14 +141,7 @@ class Writer
     {
         /** @var OpenAPISpecWriter $writer */
         $writer = app()->makeWith(OpenAPISpecWriter::class, ['config' => $this->config]);
-
         $spec = $writer->generateSpecContent($groupedEndpoints);
-        $overrides = $this->config->get('openapi.overrides', []);
-        if (count($overrides)) {
-            foreach ($overrides as $key => $value) {
-                data_set($spec, $key, $value);
-            }
-        }
         return Yaml::dump($spec, 20, 2, Yaml::DUMP_EMPTY_ARRAY_AS_SEQUENCE | Yaml::DUMP_OBJECT_AS_MAP);
     }
 
@@ -183,12 +169,18 @@ class Writer
         // Rewrite asset links to go through Laravel
         $contents = preg_replace('#href="\.\./docs/css/(.+?)"#', 'href="{{ asset("' . $this->laravelAssetsPath . '/css/$1") }}"', $contents);
         $contents = preg_replace('#src="\.\./docs/(js|images)/(.+?)"#', 'src="{{ asset("' . $this->laravelAssetsPath . '/$1/$2") }}"', $contents);
-        $contents = str_replace('href="../docs/collection.json"', 'href="{{ route("' . $this->docsName . '.postman") }}"', $contents);
-        $contents = str_replace('href="../docs/openapi.yaml"', 'href="{{ route("' . $this->docsName . '.openapi") }}"', $contents);
+        $contents = str_replace('href="../docs/collection.json"', 'href="{{ route("' . $this->paths->outputPath('postman', '.') . '") }}"', $contents);
+        $contents = str_replace('href="../docs/openapi.yaml"', 'href="{{ route("' . $this->paths->outputPath('openapi', '.') . '") }}"', $contents);
+        $contents = str_replace('url="../docs/openapi.yaml"', 'url="{{ route("' . $this->paths->outputPath('openapi', '.') . '") }}"', $contents);
+        // With Elements theme, we'd have <elements-api apiDescriptionUrl="../docs/openapi.yaml"
+        $contents = str_replace('Url="../docs/openapi.yaml"', 'Url="{{ route("' . $this->paths->outputPath('openapi', '.') . '") }}"', $contents);
 
         file_put_contents("$this->laravelTypeOutputPath/index.blade.php", $contents);
     }
 
+    /**
+     * @param array[] $groupedEndpoints
+     */
     public function writeHtmlDocs(array $groupedEndpoints): void
     {
         c::info('Writing ' . ($this->isStatic ? 'HTML' : 'Blade') . ' docs...');
@@ -196,7 +188,7 @@ class Writer
         // Then we convert them to HTML, and throw in the endpoints as well.
         /** @var HtmlWriter $writer */
         $writer = app()->makeWith(HtmlWriter::class, ['config' => $this->config]);
-        $writer->generate($groupedEndpoints, $this->markdownOutputPath, $this->staticTypeOutputPath);
+        $writer->generate($groupedEndpoints, $this->paths->intermediateOutputPath(), $this->staticTypeOutputPath);
 
         if (!$this->isStatic) {
             $this->performFinalTasksForLaravelType();
@@ -219,6 +211,31 @@ class Writer
         $this->generatedFiles['assets']['images'] = realpath("{$assetsOutputPath}images");
     }
 
+    public function writeExternalHtmlDocs(): void
+    {
+        c::info('Writing client-side HTML docs...');
+
+        /** @var ExternalHtmlWriter $writer */
+        $writer = app()->makeWith(ExternalHtmlWriter::class, ['config' => $this->config]);
+        $writer->generate([], $this->paths->intermediateOutputPath(), $this->staticTypeOutputPath);
+
+       if (!$this->isStatic) {
+           $this->performFinalTasksForLaravelType();
+       }
+
+        if ($this->isStatic) {
+            $outputPath = rtrim($this->staticTypeOutputPath, '/') . '/';
+            c::success("Wrote client-side HTML docs and assets to: $outputPath");
+            $this->generatedFiles['html'] = realpath("{$outputPath}index.html");
+        } else {
+            $outputPath = rtrim($this->laravelTypeOutputPath, '/') . '/';
+            c::success("Wrote Blade docs to: " . $this->makePathFriendly($outputPath));
+            $this->generatedFiles['blade'] = realpath("{$outputPath}index.blade.php");
+            $assetsOutputPath = public_path() . $this->laravelAssetsPath . '/';
+            c::success("Wrote Laravel assets to: " . $this->makePathFriendly($assetsOutputPath));
+        }
+    }
+
     protected function runAfterGeneratingHook()
     {
         if (is_callable(Globals::$__afterGenerating)) {
@@ -231,7 +248,10 @@ class Writer
     {
         if ($this->isStatic) return null;
 
-        return config('view.paths.0', function_exists('base_path') ? base_path("resources/views") : "resources/views") . "/$this->docsName";
+        return config(
+            'view.paths.0',
+            function_exists('base_path') ? base_path("resources/views") : "resources/views"
+        ). "/" . $this->paths->outputPath();
     }
 
     /**
