@@ -6,7 +6,7 @@ use Illuminate\Foundation\Http\FormRequest;
 
 class TestRequestQueryParams extends FormRequest
 {
-    public function rules()
+    public function rules(\stdClass $injectedDep)
     {
         return [
             'q_param' => 'int|required',
