@@ -2,7 +2,6 @@
 
 namespace Knuckles\Scribe\Tests\Strategies\Responses;
 
-use Dingo\Api\Routing\Router;
 use Illuminate\Routing\Route;
 use Illuminate\Support\Facades\Route as RouteFacade;
 use Knuckles\Camel\Extraction\ExtractedEndpointData;
@@ -13,46 +12,26 @@ use Knuckles\Scribe\Scribe;
 use Knuckles\Scribe\Tests\BaseLaravelTest;
 use Knuckles\Scribe\Tests\Fixtures\TestController;
 use Knuckles\Scribe\Tools\DocumentationConfig;
-use DMS\PHPUnitExtensions\ArraySubset\ArraySubsetAsserts;
 use Illuminate\Support\Facades\Route as LaravelRouteFacade;
 use Symfony\Component\HttpFoundation\Request;
 
 class ResponseCallsTest extends BaseLaravelTest
 {
-    use ArraySubsetAsserts;
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-        config(['scribe.database_connections_to_transact' => []]);
-    }
-
     /** @test */
     public function can_call_route_and_fetch_response()
     {
         $route = LaravelRouteFacade::post('/shouldFetchRouteResponse', [TestController::class, 'shouldFetchRouteResponse']);
 
-        $rules = [
-            'headers' => [
-                'Content-Type' => 'application/json',
-                'Accept' => 'application/json',
-            ],
-            'response_calls' => [
-                'methods' => ['*'],
-            ],
-        ];
+        $responses = $this->invokeStrategy($route);
 
-        $strategy = new ResponseCalls(new DocumentationConfig([]));
-        $results = $strategy(ExtractedEndpointData::fromRoute($route), $rules);
-
-        $this->assertEquals(200, $results[0]['status']);
+        $this->assertEquals(200, $responses[0]['status']);
         $this->assertArraySubset([
             'id' => 4,
             'name' => 'banana',
             'color' => 'red',
             'weight' => '1 kg',
             'delicious' => true,
-        ], json_decode($results[0]['content'], true));
+        ], json_decode($responses[0]['content'], true));
     }
 
     /** @test */
@@ -60,10 +39,23 @@ class ResponseCallsTest extends BaseLaravelTest
     {
         $route = RouteFacade::post('/withFormDataParams', [TestController::class, 'withFormDataParams']);
 
-        config(['scribe.routes.0.apply.response_calls.methods' => ['POST']]);
-        $parsed = (new Extractor())->processRoute($route, config('scribe.routes.0.apply'));
-
+        /* This doesn't work. It always gives an error, "the file failed to upload". However, uploading files worked when they are extracted params
+        $endpointData = ExtractedEndpointData::fromRoute($route, [ 'headers' => [ 'accept' => 'application/json' ] ]);
+        $responses = $this->invokeStrategy($endpointData, settings: [
+            'fileParams' => [ 'image' => 'config/scribe.php' ],
+            'bodyParams' => [ 'name' => 'cat.jpg' ]
+        ]);
+        */
+        $this->setConfig([
+            'strategies.responses' => [
+                [ResponseCalls::class,
+                    ['only' => 'POST *']
+                ],
+            ]
+        ]);
+        $parsed = (new Extractor())->processRoute($route);
         $responses = $parsed->responses->toArray();
+
         $this->assertCount(1, $responses);
         $this->assertArraySubset([
             "status" => 200,
@@ -77,18 +69,6 @@ class ResponseCallsTest extends BaseLaravelTest
     {
         $route = LaravelRouteFacade::post('/echo/{id}', [TestController::class, 'echoesRequestValues']);
 
-        $rules = [
-            'response_calls' => [
-                'methods' => ['*'],
-                'queryParams' => [
-                    'queryParam' => 'queryValue',
-                ],
-                'bodyParams' => [
-                    'bodyParam' => 'bodyValue',
-                ],
-            ],
-        ];
-
         $endpointData = ExtractedEndpointData::fromRoute($route, [
             'auth' => ['headers', 'Authorization', 'Bearer bearerToken'],
             'headers' => [
@@ -98,12 +78,18 @@ class ResponseCallsTest extends BaseLaravelTest
             ],
         ]);
 
-        $strategy = new ResponseCalls(new DocumentationConfig([]));
-        $results = $strategy($endpointData, $rules);
+        $responses = $this->invokeStrategy($endpointData, settings: [
+            'queryParams' => [
+                'queryParam' => 'queryValue',
+            ],
+            'bodyParams' => [
+                'bodyParam' => 'bodyValue',
+            ],
+        ]);
 
-        $this->assertEquals(200, $results[0]['status']);
+        $this->assertEquals(200, $responses[0]['status']);
 
-        $responseContent = json_decode($results[0]['content'], true);
+        $responseContent = json_decode($responses[0]['content'], true);
         $this->assertEquals('queryValue', $responseContent['queryParam']);
         $this->assertEquals('bodyValue', $responseContent['bodyParam']);
         $this->assertEquals('headerValue', $responseContent['header']);
@@ -114,29 +100,16 @@ class ResponseCallsTest extends BaseLaravelTest
     public function can_override_application_config_during_response_call()
     {
         $route = LaravelRouteFacade::post('/echoesConfig', [TestController::class, 'echoesConfig']);
-
-        $rules = [
-            'response_calls' => [
-                'methods' => ['*'],
-            ],
-        ];
-
-        $strategy = new ResponseCalls(new DocumentationConfig([]));
-        $results = $strategy(ExtractedEndpointData::fromRoute($route), $rules);
-        $originalValue = json_decode($results[0]['content'], true)['app.env'];
+        $responses = $this->invokeStrategy($route);
+        $originalValue = json_decode($responses[0]['content'], true)['app.env'];
 
         $now = time();
-        $rules = [
-            'response_calls' => [
-                'methods' => ['*'],
-                'config' => [
-                    'app.env' => $now,
-                ],
+        $responses = $this->invokeStrategy($route, settings: [
+            'config' => [
+                'app.env' => $now,
             ],
-        ];
-
-        $results = $strategy(ExtractedEndpointData::fromRoute($route), $rules);
-        $newValue = json_decode($results[0]['content'], true)['app.env'];
+        ],);
+        $newValue = json_decode($responses[0]['content'], true)['app.env'];
         $this->assertEquals($now, $newValue);
         $this->assertNotEquals($originalValue, $newValue);
     }
@@ -153,18 +126,6 @@ class ResponseCallsTest extends BaseLaravelTest
 
         $route = LaravelRouteFacade::post('/echo/{id}', [TestController::class, 'echoesRequestValues']);
 
-        $rules = [
-            'response_calls' => [
-                'methods' => ['*'],
-                'queryParams' => [
-                    'queryParam' => 'queryValue',
-                ],
-                'bodyParams' => [
-                    'bodyParam' => 'bodyValue',
-                ],
-            ],
-        ];
-
         $endpointData = ExtractedEndpointData::fromRoute($route, [
             'auth' => ['headers', 'Authorization', 'Bearer bearerToken'],
             'headers' => [
@@ -173,13 +134,18 @@ class ResponseCallsTest extends BaseLaravelTest
                 'header' => 'headerValue',
             ],
         ]);
+        $responses = $this->invokeStrategy($endpointData, settings: [
+            'queryParams' => [
+                'queryParam' => 'queryValue',
+            ],
+            'bodyParams' => [
+                'bodyParam' => 'bodyValue',
+            ],
+        ]);
 
-        $strategy = new ResponseCalls(new DocumentationConfig([]));
-        $results = $strategy($endpointData, $rules);
+        $this->assertEquals(200, $responses[0]['status']);
 
-        $this->assertEquals(200, $results[0]['status']);
-
-        $responseContent = json_decode($results[0]['content'], true);
+        $responseContent = json_decode($responses[0]['content'], true);
         $this->assertEquals('overridden_queryValue', $responseContent['queryParam']);
         $this->assertEquals('overridden_headerValue', $responseContent['header']);
         $this->assertEquals('overridden_Bearer bearerToken', $responseContent['auth']);
@@ -188,123 +154,10 @@ class ResponseCallsTest extends BaseLaravelTest
         Scribe::beforeResponseCall(fn() => null);
     }
 
-    /**
-     * @test
-     * @group dingo
-     */
-    public function can_call_route_and_fetch_response_with_dingo()
-    {
-        $route = $this->registerDingoRoute('post', '/shouldFetchRouteResponse', 'shouldFetchRouteResponse');
-
-        $rules = [
-            'headers' => [
-                'Content-Type' => 'application/json',
-                'Accept' => 'application/json',
-            ],
-            'response_calls' => [
-                'methods' => ['*'],
-            ],
-        ];
-
-        $strategy = new ResponseCalls(new DocumentationConfig());
-        $results = $strategy(ExtractedEndpointData::fromRoute($route), $rules);
-
-        $this->assertEquals(200, $results[0]['status']);
-        $this->assertArraySubset([
-            'id' => 4,
-            'name' => 'banana',
-            'color' => 'red',
-            'weight' => '1 kg',
-            'delicious' => true,
-        ], json_decode($results[0]['content'], true));
-    }
-
-    /**
-     * @test
-     * @group dingo
-     */
-    public function uses_configured_settings_when_calling_route_with_dingo()
-    {
-        $route = $this->registerDingoRoute('post', '/echo/{id}', 'echoesRequestValues');
-
-        $rules = [
-            'response_calls' => [
-                'methods' => ['*'],
-                'queryParams' => [
-                    'queryParam' => 'queryValue',
-                ],
-                'bodyParams' => [
-                    'bodyParam' => 'bodyValue',
-                ],
-            ],
-        ];
-
-        $endpointData = ExtractedEndpointData::fromRoute($route, [
-            'headers' => [
-                'Content-Type' => 'application/json',
-                'Accept' => 'application/json',
-                'header' => 'headerValue',
-            ],
-        ]);
-        $strategy = new ResponseCalls(new DocumentationConfig());
-        $results = $strategy($endpointData, $rules);
-
-        $this->assertEquals(200, $results[0]['status']);
-
-        $responseContent = json_decode($results[0]['content'], true);
-        $this->assertEquals('queryValue', $responseContent['queryParam']);
-        $this->assertEquals('bodyValue', $responseContent['bodyParam']);
-        $this->assertEquals('headerValue', $responseContent['header']);
-    }
-
-    /**
-     * @test
-     * @group dingo
-     */
-    public function can_override_application_config_during_response_call_with_dingo()
-    {
-        $route = $this->registerDingoRoute('post', '/echoesConfig', 'echoesConfig');
-
-        $rules = [
-            'response_calls' => [
-                'methods' => ['*'],
-            ],
-        ];
-
-        $strategy = new ResponseCalls(new DocumentationConfig());
-        $results = $strategy(ExtractedEndpointData::fromRoute($route), $rules);
-        $originalValue = json_decode($results[0]['content'], true)['app.env'];
-
-        $now = time();
-        $rules = [
-            'response_calls' => [
-                'methods' => ['*'],
-                'config' => [
-                    'app.env' => $now,
-                ],
-            ],
-        ];
-
-        $results = $strategy(ExtractedEndpointData::fromRoute($route), $rules);
-        $newValue = json_decode($results[0]['content'], true)['app.env'];
-        $this->assertEquals($now, $newValue);
-        $this->assertNotEquals($originalValue, $newValue);
-    }
-
     /** @test */
     public function does_not_make_response_call_if_success_response_already_gotten()
     {
         $route = LaravelRouteFacade::post('/shouldFetchRouteResponse', [TestController::class, 'shouldFetchRouteResponse']);
-
-        $rules = [
-            'headers' => [
-                'Content-Type' => 'application/json',
-                'Accept' => 'application/json',
-            ],
-            'response_calls' => [
-                'methods' => ['*'],
-            ],
-        ];
 
         $endpointData = ExtractedEndpointData::fromRoute($route, [
             'responses' => new ResponseCollection([
@@ -314,44 +167,38 @@ class ResponseCallsTest extends BaseLaravelTest
                 ],
             ]),
         ]);
-        $strategy = new ResponseCalls(new DocumentationConfig([]));
-        $results = $strategy($endpointData, $rules);
+        $responses = $this->invokeStrategy($endpointData);
 
-        $this->assertNull($results);
+        $this->assertNull($responses);
     }
 
     /** @test */
-    public function does_not_make_response_call_if_forbidden_by_config()
+    public function can_get_content_from_streamed_response()
     {
-        $route = LaravelRouteFacade::post('/shouldFetchRouteResponse', [TestController::class, 'shouldFetchRouteResponse']);
+        $route = LaravelRouteFacade::post('/withStreamedResponse', [TestController::class, 'withStreamedResponse']);
 
-        $rules = [
-            'response_calls' => [
-                'methods' => [],
-            ],
-        ];
-        $strategy = new ResponseCalls(new DocumentationConfig([]));
-        $results = $strategy(ExtractedEndpointData::fromRoute($route), $rules);
+        $this->withoutExceptionHandling();
+        $responses = $this->invokeStrategy($route);
 
-        $this->assertNull($results);
+        $this->assertEquals(200, $responses[0]['status']);
+        $this->assertArraySubset([
+            'items' => [
+                'one',
+                'two',
+            ]
+        ], json_decode($responses[0]['content'], true));
     }
 
-    public function registerDingoRoute(string $httpMethod, string $path, string $controllerMethod)
+    protected function convertRules(array $rules): mixed
     {
-        $desiredRoute = null;
-        /** @var Router $api */
-        $api = app(Router::class);
-        $api->version('v1', function (Router $api) use ($controllerMethod, $path, $httpMethod, &$desiredRoute) {
-            $desiredRoute = $api->$httpMethod($path, [TestController::class, $controllerMethod]);
-        });
-        $routes = app(\Dingo\Api\Routing\Router::class)->getRoutes('v1');
+        return Extractor::transformOldRouteRulesIntoNewSettings('responses', $rules, ResponseCalls::class);
+    }
 
-        /*
-         * Doing this bc we want an instance of Dingo\Api\Routing\Route, not Illuminate\Routing\Route, which the method above returns
-         */
-        return collect($routes)
-            ->first(function (Route $route) use ($desiredRoute) {
-                return $route->uri() === $desiredRoute->uri();
-            });
+    protected function invokeStrategy(ExtractedEndpointData|Route $route, $settings = []): ?array
+    {
+        $strategy = new ResponseCalls(new DocumentationConfig([]));
+        return $strategy(
+            $route instanceof ExtractedEndpointData ? $route : ExtractedEndpointData::fromRoute($route), $settings
+        );
     }
 }
