@@ -2,7 +2,10 @@
 
 namespace Knuckles\Scribe\Tests\Strategies\Responses;
 
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Foundation\Application;
 use Illuminate\Routing\Route;
+use Illuminate\Support\Facades\Schema;
 use Knuckles\Camel\Extraction\ExtractedEndpointData;
 use Knuckles\Scribe\Attributes\Response;
 use Knuckles\Scribe\Attributes\ResponseFromApiResource;
@@ -19,13 +22,10 @@ use Knuckles\Scribe\Tests\Fixtures\TestUserApiResource;
 use Knuckles\Scribe\Tools\DocumentationConfig;
 use Knuckles\Scribe\Tools\Utils;
 use League\Fractal\Pagination\IlluminatePaginatorAdapter;
-use DMS\PHPUnitExtensions\ArraySubset\ArraySubsetAsserts;
 use ReflectionClass;
 
 class UseResponseAttributesTest extends BaseLaravelTest
 {
-    use ArraySubsetAsserts;
-
     protected function getPackageProviders($app)
     {
         $providers = parent::getPackageProviders($app);
@@ -39,7 +39,7 @@ class UseResponseAttributesTest extends BaseLaravelTest
     {
         parent::setUp();
 
-        config(['scribe.database_connections_to_transact' => []]);
+        $this->setConfig(['database_connections_to_transact' => []]);
 
         $factory = app(\Illuminate\Database\Eloquent\Factory::class);
         $factory->define(TestUser::class, function () {
@@ -76,6 +76,10 @@ class UseResponseAttributesTest extends BaseLaravelTest
                 'status' => 201,
                 'content' => json_encode(["all" => "good"]),
             ],
+            [
+                'status' => 404,
+                'content' => null,
+            ]
         ], $results);
     }
 
@@ -227,7 +231,141 @@ class UseResponseAttributesTest extends BaseLaravelTest
         ], $results);
     }
 
-    protected function fetch($endpoint): array
+    /** @test */
+    public function can_parse_apiresource_attributes_with_cursor_pagination()
+    {
+        $factory = app(\Illuminate\Database\Eloquent\Factory::class);
+        $factory->afterMaking(TestUser::class, function (TestUser $user, $faker) {
+            if ($user->id === 4) {
+                $child = Utils::getModelFactory(TestUser::class)->make(['id' => 5, 'parent_id' => 4]);
+                $user->setRelation('children', collect([$child]));
+            }
+        });
+
+        $results = $this->fetch($this->endpoint("apiResourceAttributesWithCursorPaginate"));
+
+
+        $nextCursor = base64_encode(json_encode(['_pointsToNextItems' => true]));
+        $this->assertArraySubset([
+            [
+                'status' => 200,
+                'content' => json_encode([
+                    'data' => [
+                        [
+                            'id' => 4,
+                            'name' => 'Tested Again',
+                            'email' => 'a@b.com',
+                            'children' => [
+                                [
+                                    'id' => 5,
+                                    'name' => 'Tested Again',
+                                    'email' => 'a@b.com',
+                                ],
+                            ],
+                        ],
+                    ],
+                    'links' => [
+                        "first" => null,
+                        "last" => null,
+                        "prev" => null,
+                        "next" => "/?cursor={$nextCursor}",
+                    ],
+                    "meta" => [
+                        "path" => '/',
+                        'per_page' => 1,
+                        'next_cursor' => $nextCursor,
+                        'prev_cursor' => null,
+                    ]
+                ]),
+            ],
+        ], $results);
+    }
+
+    /** @test */
+    public function can_parse_apiresource_attributes_and_load_children_using_factory_create()
+    {
+        Schema::create('test_users', function (Blueprint $table) {
+            $table->id();
+            $table->string('first_name');
+            $table->string('last_name');
+            $table->string('email');
+            $table->integer('parent_id')->nullable();
+        });
+
+        $factory = app(\Illuminate\Database\Eloquent\Factory::class);
+        $factory->afterCreating(TestUser::class, function (TestUser $user, $faker) {
+            if ($user->id === 4) {
+                Utils::getModelFactory(TestUser::class)->create(['id' => 5, 'parent_id' => 4]);
+            }
+        });
+        $documentationConfig = ['examples' => ['models_source' => ['factoryCreate']]];
+
+        $results = $this->fetch($this->endpoint("apiResourceAttributesIncludeChildren"), $documentationConfig);
+        $this->assertArraySubset([
+            [
+                'status' => 200,
+                'content' => json_encode([
+                    "data" => [
+                        "id" => 4,
+                        "name" => "Tested Again",
+                        "email" => "a@b.com",
+                        "children" => [
+                            [
+                                "id" => 5,
+                                "name" => "Tested Again",
+                                "email" => "a@b.com",
+                            ]
+                        ],
+                    ],
+                ]),
+            ],
+        ], $results);
+    }
+
+
+    /** @test */
+    public function can_parse_apiresource_attributes_and_load_children_and_children_count_using_factory_create()
+    {
+        Schema::create('test_users', function (Blueprint $table) {
+            $table->id();
+            $table->string('first_name');
+            $table->string('last_name');
+            $table->string('email');
+            $table->integer('parent_id')->nullable();
+        });
+
+        $factory = app(\Illuminate\Database\Eloquent\Factory::class);
+        $factory->afterCreating(TestUser::class, function (TestUser $user, $faker) {
+            if ($user->id === 4) {
+                Utils::getModelFactory(TestUser::class)->create(['id' => 5, 'parent_id' => 4]);
+            }
+        });
+        $documentationConfig = ['examples' => ['models_source' => ['factoryCreate']]];
+
+        $results = $this->fetch($this->endpoint("apiResourceAttributesIncludeChildrenAndChildrenCount"), $documentationConfig);
+        $this->assertArraySubset([
+            [
+                'status' => 200,
+                'content' => json_encode([
+                    "data" => [
+                        "id" => 4,
+                        "name" => "Tested Again",
+                        "email" => "a@b.com",
+                        "children" => [
+                            [
+                                "id" => 5,
+                                "name" => "Tested Again",
+                                "email" => "a@b.com",
+                            ]
+                        ],
+                        'children_count' => 1,
+                    ],
+                ]),
+            ],
+        ], $results);
+    }
+
+    protected function fetch($endpoint, array $documentationConfig = []): array
     {
         $strategy = new UseResponseAttributes(new DocumentationConfig([]));
         return $strategy($endpoint, []);
@@ -249,6 +387,7 @@ class ResponseAttributesTestController
 {
     #[Response(["all" => "good"], 200, "Success")]
     #[Response('{"all":"good"}', 201)]
+    #[Response(status: 404)]
     public function plainResponseAttributes()
     {
 
@@ -277,6 +416,24 @@ class ResponseAttributesTestController
     #[ResponseFromTransformer(TestTransformer::class, TestModel::class, collection: true,
         paginate: [IlluminatePaginatorAdapter::class, 1])]
     public function transformerAttributes()
+    {
+
+    }
+
+    #[ResponseFromApiResource(TestUserApiResource::class, collection: true, cursorPaginate: 1)]
+    public function apiResourceAttributesWithCursorPaginate()
+    {
+
+    }
+
+    #[ResponseFromApiResource(TestUserApiResource::class, with: ['children'], withCount: ['children'])]
+    public function apiResourceAttributesIncludeChildrenAndChildrenCount()
+    {
+
+    }
+
+    #[ResponseFromApiResource(TestUserApiResource::class, with: ['children'])]
+    public function apiResourceAttributesIncludeChildren()
     {
 
     }
