@@ -2,22 +2,21 @@
 
 namespace Knuckles\Scribe\Tests\Unit;
 
-use DMS\PHPUnitExtensions\ArraySubset\ArraySubsetAsserts;
 use Faker\Factory;
 use Illuminate\Support\Arr;
 use Knuckles\Camel\Camel;
 use Knuckles\Camel\Output\OutputEndpointData;
+use Knuckles\Scribe\Tests\BaseUnitTest;
+use Knuckles\Scribe\Tests\Fixtures\ComponentsOpenApiGenerator;
+use Knuckles\Scribe\Tests\Fixtures\TestOpenApiGenerator;
 use Knuckles\Scribe\Tools\DocumentationConfig;
 use Knuckles\Scribe\Writing\OpenAPISpecWriter;
-use PHPUnit\Framework\TestCase;
 
 /**
  * See https://swagger.io/specification/
  */
-class OpenAPISpecWriterTest extends TestCase
+class OpenAPISpecWriterTest extends BaseUnitTest
 {
-    use ArraySubsetAsserts;
-
     protected $config = [
         'title' => 'My Testy Testes API',
         'description' => 'All about testy testes.',
@@ -80,8 +79,14 @@ class OpenAPISpecWriterTest extends TestCase
         $endpointData1 = $this->createMockEndpointData(['uri' => 'path1', 'httpMethods' => ['GET'], 'metadata.authenticated' => true]);
         $endpointData2 = $this->createMockEndpointData(['uri' => 'path1', 'httpMethods' => ['POST'], 'metadata.authenticated' => false]);
         $groups = [$this->createGroup([$endpointData1, $endpointData2])];
-
-        $config = array_merge($this->config, ['auth' => ['enabled' => true, 'in' => 'bearer']]);
+        $extraInfo = "When stuck trying to authenticate, have a coffee!";
+        $config = array_merge($this->config, [
+            'auth' => [
+                'enabled' => true,
+                'in' => 'bearer',
+                'extra_info' => $extraInfo,
+            ],
+        ]);
         $writer = new OpenAPISpecWriter(new DocumentationConfig($config));
         $results = $writer->generateSpecContent($groups);
 
@@ -89,6 +94,7 @@ class OpenAPISpecWriterTest extends TestCase
         $this->assertArrayHasKey('default', $results['components']['securitySchemes']);
         $this->assertEquals('http', $results['components']['securitySchemes']['default']['type']);
         $this->assertEquals('bearer', $results['components']['securitySchemes']['default']['scheme']);
+        $this->assertEquals($extraInfo, $results['components']['securitySchemes']['default']['description']);
         $this->assertCount(1, $results['security']);
         $this->assertCount(1, $results['security'][0]);
         $this->assertArrayHasKey('default', $results['security'][0]);
@@ -97,13 +103,21 @@ class OpenAPISpecWriterTest extends TestCase
         $this->assertCount(0, $results['paths']['/path1']['post']['security']);
 
         // Next try: auth with a query parameter
-        $config = array_merge($this->config, ['auth' => ['enabled' => true, 'in' => 'query', 'name' => 'token']]);
+        $config = array_merge($this->config, [
+            'auth' => [
+                'enabled' => true,
+                'in' => 'query',
+                'name' => 'token',
+                'extra_info' => $extraInfo,
+            ],
+        ]);
         $writer = new OpenAPISpecWriter(new DocumentationConfig($config));
         $results = $writer->generateSpecContent($groups);
 
         $this->assertCount(1, $results['components']['securitySchemes']);
         $this->assertArrayHasKey('default', $results['components']['securitySchemes']);
         $this->assertEquals('apiKey', $results['components']['securitySchemes']['default']['type']);
+        $this->assertEquals($extraInfo, $results['components']['securitySchemes']['default']['description']);
         $this->assertEquals($config['auth']['name'], $results['components']['securitySchemes']['default']['name']);
         $this->assertEquals('query', $results['components']['securitySchemes']['default']['in']);
         $this->assertCount(1, $results['security']);
@@ -112,6 +126,22 @@ class OpenAPISpecWriterTest extends TestCase
         $this->assertArrayNotHasKey('security', $results['paths']['/path1']['get']);
         $this->assertArrayHasKey('security', $results['paths']['/path1']['post']);
         $this->assertCount(0, $results['paths']['/path1']['post']['security']);
+    }
+
+    /** @test */
+    public function adds_deprecation_info_correctly()
+    {
+        $endpointData1 = $this->createMockEndpointData(['uri' => 'path1', 'httpMethods' => ['GET'], 'metadata.deprecated' => true]);
+        $endpointData2 = $this->createMockEndpointData(['uri' => 'path2', 'httpMethods' => ['GET'], 'metadata.deprecated' => false]);
+        $groups = [$this->createGroup([$endpointData1, $endpointData2])];
+
+        $results = $this->generate($groups);
+
+        $this->assertIsArray($results['paths']);
+        $this->assertCount(2, $results['paths']);
+        $this->assertArrayHasKey('deprecated', $results['paths']['/path1']['get']);
+        $this->assertTrue(true, $results['paths']['/path1']['get']['deprecated']);
+        $this->assertArrayNotHasKey('deprecated', $results['paths']['/path2']['get']);
     }
 
     /** @test */
@@ -198,6 +228,7 @@ class OpenAPISpecWriterTest extends TestCase
                     'example' => 'hahoho',
                     'type' => 'string',
                     'name' => 'param',
+                    'nullable' => false
                 ],
             ],
         ]);
@@ -219,6 +250,7 @@ class OpenAPISpecWriterTest extends TestCase
                 'type' => 'string',
                 'description' => 'A query param',
                 'example' => 'hahoho',
+                'nullable' => false
             ],
         ], $results['paths']['/path1']['get']['parameters'][0]);
     }
@@ -236,6 +268,7 @@ class OpenAPISpecWriterTest extends TestCase
                     'required' => false,
                     'example' => 'hahoho',
                     'type' => 'string',
+                    'nullable' => false,
                 ],
                 'integerParam' => [
                     'name' => 'integerParam',
@@ -243,6 +276,7 @@ class OpenAPISpecWriterTest extends TestCase
                     'required' => true,
                     'example' => 99,
                     'type' => 'integer',
+                    'nullable' => false,
                 ],
                 'booleanParam' => [
                     'name' => 'booleanParam',
@@ -250,6 +284,7 @@ class OpenAPISpecWriterTest extends TestCase
                     'required' => true,
                     'example' => false,
                     'type' => 'boolean',
+                    'nullable' => false,
                 ],
                 'objectParam' => [
                     'name' => 'objectParam',
@@ -257,6 +292,7 @@ class OpenAPISpecWriterTest extends TestCase
                     'required' => false,
                     'example' => [],
                     'type' => 'object',
+                    'nullable' => false,
                 ],
                 'objectParam.field' => [
                     'name' => 'objectParam.field',
@@ -264,6 +300,7 @@ class OpenAPISpecWriterTest extends TestCase
                     'required' => false,
                     'example' => 119.0,
                     'type' => 'number',
+                    'nullable' => false,
                 ],
             ],
         ]);
@@ -326,26 +363,31 @@ class OpenAPISpecWriterTest extends TestCase
                                 'description' => 'String param',
                                 'example' => 'hahoho',
                                 'type' => 'string',
+                                'nullable' => false,
                             ],
                             'booleanParam' => [
                                 'description' => 'Boolean param',
                                 'example' => false,
                                 'type' => 'boolean',
+                                'nullable' => false,
                             ],
                             'integerParam' => [
                                 'description' => 'Integer param',
                                 'example' => 99,
                                 'type' => 'integer',
+                                'nullable' => false,
                             ],
                             'objectParam' => [
                                 'description' => 'Object param',
                                 'example' => [],
                                 'type' => 'object',
+                                'nullable' => false,
                                 'properties' => [
                                     'field' => [
                                         'description' => 'Object param field',
                                         'example' => 119.0,
                                         'type' => 'number',
+                                        'nullable' => false,
                                     ],
                                 ],
                             ],
@@ -369,6 +411,7 @@ class OpenAPISpecWriterTest extends TestCase
                                 'description' => 'File param',
                                 'type' => 'string',
                                 'format' => 'binary',
+                                'nullable' => false,
                             ],
                             'numberArrayParam' => [
                                 'description' => 'Number array param',
@@ -398,6 +441,7 @@ class OpenAPISpecWriterTest extends TestCase
                                             'type' => 'string',
                                             'description' => '',
                                             'example' => "hi",
+                                            'nullable' => false,
                                         ],
                                     ],
                                 ],
@@ -424,7 +468,7 @@ class OpenAPISpecWriterTest extends TestCase
                 [
                     'status' => 201,
                     'description' => '',
-                    'content' => '{"this": "shouldn\'t be ignored", "and this": "too", "sub level 0": { "sub level 1 key 1": "sl0_sl1k1", "sub level 1 key 2": [ { "sub level 2 key 1": "sl0_sl1k2_sl2k1", "sub level 2 key 2": { "sub level 3 key 1": "sl0_sl1k2_sl2k2_sl3k1" } } ], "sub level 1 key 3": { "sub level 2 key 1": "sl0_sl1k3_sl2k2", "sub level 2 key 2": { "sub level 3 key 1": "sl0_sl1k3_sl2k2_sl3k1", "sub level 3 key null": null, "sub level 3 key integer": 99 } } } }',
+                    'content' => '{"this": "shouldn\'t be ignored", "and this": "too", "also this": "too", "sub level 0": { "sub level 1 key 1": "sl0_sl1k1", "sub level 1 key 2": [ { "sub level 2 key 1": "sl0_sl1k2_sl2k1", "sub level 2 key 2": { "sub level 3 key 1": "sl0_sl1k2_sl2k2_sl3k1" } } ], "sub level 1 key 3": { "sub level 2 key 1": "sl0_sl1k3_sl2k2", "sub level 2 key 2": { "sub level 3 key 1": "sl0_sl1k3_sl2k2_sl3k1", "sub level 3 key null": null, "sub level 3 key integer": 99 }, "sub level 2 key 3 required" : "sl0_sl1k3_sl2k3" } } }',
                 ],
             ],
             'responseFields' => [
@@ -433,9 +477,19 @@ class OpenAPISpecWriterTest extends TestCase
                     'type' => 'string',
                     'description' => 'Parameter description, ha!',
                 ],
-                'sub level 0.sub level 1 key 3.sub level 2 key 1'=> [
-                    'description' => 'This is description of nested object',
-                ]
+                'also this' => [
+                    'name' => 'also this',
+                    'type' => 'string',
+                    'description' => 'This response parameter is required.',
+                    'required' => true,
+                ],
+                'sub level 0.sub level 1 key 3.sub level 2 key 1' => [
+                    'description' => 'This is a description of a nested object',
+                ],
+                'sub level 0.sub level 1 key 3.sub level 2 key 3 required' => [
+                    'description' => 'This is a description of a required nested object',
+                    'required' => true,
+                ],
             ],
         ]);
         $endpointData2 = $this->createMockEndpointData([
@@ -473,6 +527,11 @@ class OpenAPISpecWriterTest extends TestCase
                                     'example' => "too",
                                     'type' => 'string',
                                 ],
+                                'also this' => [
+                                    'description' => 'This response parameter is required.',
+                                    'example' => "too",
+                                    'type' => 'string',
+                                ],
                                 'sub level 0' => [
                                     'type' => 'object',
                                     'properties' => [
@@ -500,7 +559,7 @@ class OpenAPISpecWriterTest extends TestCase
                                                 'sub level 2 key 1' => [
                                                     'type' => 'string',
                                                     'example' => 'sl0_sl1k3_sl2k2',
-                                                    'description' => 'This is description of nested object'
+                                                    'description' => 'This is a description of a nested object'
                                                 ],
                                                 'sub level 2 key 2' => [
                                                     'type' => 'object',
@@ -518,12 +577,24 @@ class OpenAPISpecWriterTest extends TestCase
                                                             'example' => 99
                                                         ]
                                                     ]
-                                                ]
+                                                ],
+                                                'sub level 2 key 3 required' => [
+                                                    'type' => 'string',
+                                                    'example' => 'sl0_sl1k3_sl2k3',
+                                                    'description' => 'This is a description of a required nested object'
+                                                ],
+
+                                            ],
+                                            'required' => [
+                                                'sub level 2 key 3 required'
                                             ]
                                         ]
                                     ]
                                 ]
                             ],
+                            'required' => [
+                                'also this'
+                            ]
                         ],
                     ],
                 ],
@@ -543,6 +614,525 @@ class OpenAPISpecWriterTest extends TestCase
                 ],
             ],
         ], $results['paths']['/path2']['put']['responses']);
+    }
+
+    /** @test */
+    public function adds_required_fields_on_array_of_objects()
+    {
+        $endpointData = $this->createMockEndpointData([
+            'httpMethods' => ['GEt'],
+            'uri' => '/path1',
+            'responses' => [
+                [
+                    'status' => 200,
+                    'description' => 'List of entities',
+                    'content' => '{"data":[{"name":"Resource name","uuid":"UUID","primary":true}]}',
+                ],
+            ],
+            'responseFields' => [
+                'data' => [
+                    'name' => 'data',
+                    'type' => 'array',
+                    'description' => 'Data wrapper',
+                ],
+                'data.name' => [
+                    'name' => 'Resource name',
+                    'type' => 'string',
+                    'description' => 'Name of the resource object',
+                    'required' => true,
+                ],
+                'data.uuid' => [
+                    'name' => 'Resource UUID',
+                    'type' => 'string',
+                    'description' => 'Unique ID for the resource',
+                    'required' => true,
+                ],
+                'data.primary' => [
+                    'name' => 'Is primary',
+                    'type' => 'bool',
+                    'description' => 'Is primary resource',
+                    'required' => true,
+                ],
+            ],
+        ]);
+
+        $groups = [$this->createGroup([$endpointData])];
+
+        $results = $this->generate($groups);
+
+        $this->assertArraySubset([
+            '200' => [
+                'description' => 'List of entities',
+                'content' => [
+                    'application/json' => [
+                        'schema' => [
+                            'type' => 'object',
+                            'properties' => [
+                                'data' => [
+                                    'type' => 'array',
+                                    'description' => 'Data wrapper',
+                                    'items' => [
+                                        'type' => 'object',
+                                        'properties' => [
+                                            'name' => [
+                                                'type' => 'string',
+                                                'description' => 'Name of the resource object',
+                                            ],
+                                            'uuid' => [
+                                                'type' => 'string',
+                                                'description' => 'Unique ID for the resource',
+                                            ],
+                                            'primary' => [
+                                                'type' => 'boolean',
+                                                'description' => 'Is primary resource',
+                                            ],
+                                        ],
+                                    ],
+                                    'required' => [
+                                        'name',
+                                        'uuid',
+                                        'primary',
+                                    ]
+                                ],
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+        ], $results['paths']['/path1']['get']['responses']);
+    }
+
+    /** @test */
+    public function generates_correctly_for_array_of_strings()
+    {
+        $endpointData = $this->createMockEndpointData([
+            'httpMethods' => ['GET'],
+            'uri' => '/path1',
+            'responses' => [
+                [
+                    'status' => 200,
+                    'description' => 'List of entities',
+                    'content' => '{"data":["Resource name"]}',
+                ],
+            ],
+            'responseFields' => [
+                'data' => [
+                    'name' => 'data',
+                    'type' => 'string[]',
+                    'description' => 'Data wrapper',
+                ],
+            ],
+        ]);
+
+        $groups = [$this->createGroup([$endpointData])];
+
+        $results = $this->generate($groups);
+
+        $this->assertArraySubset([
+            '200' => [
+                'description' => 'List of entities',
+                'content' => [
+                    'application/json' => [
+                        'schema' => [
+                            'type' => 'object',
+                            'properties' => [
+                                'data' => [
+                                    'type' => 'array',
+                                    'description' => 'Data wrapper',
+                                    'items' => [
+                                        'type' => 'string',
+                                    ],
+                                ],
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+        ], $results['paths']['/path1']['get']['responses']);
+    }
+
+    /** @test */
+    public function adds_multiple_responses_correctly_using_oneOf()
+    {
+        $endpointData1 = $this->createMockEndpointData([
+            'httpMethods' => ['POST'],
+            'uri' => '/path1',
+            'responses' => [
+                [
+                    'status' => 201,
+                    'description' => 'This one',
+                    'content' => '{"this": "one"}',
+                ],
+                [
+                    'status' => 201,
+                    'description' => 'No, that one.',
+                    'content' => '{"that": "one"}',
+                ],
+                [
+                    'status' => 200,
+                    'description' => 'A separate one',
+                    'content' => '{"the other": "one"}',
+                ],
+            ],
+        ]);
+        $groups = [$this->createGroup([$endpointData1])];
+
+        $results = $this->generate($groups);
+
+        $this->assertArraySubset([
+            '200' => [
+                'description' => 'A separate one',
+                'content' => [
+                    'application/json' => [
+                        'schema' => [
+                            'type' => 'object',
+                            'properties' => [
+                                'the other' => [
+                                    'example' => "one",
+                                    'type' => 'string',
+                                ],
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+            '201' => [
+                'description' => '',
+                'content' => [
+                    'application/json' => [
+                        'schema' => [
+                            'oneOf' => [
+                                [
+                                    'type' => 'object',
+                                    'description' => 'This one',
+                                    'properties' => [
+                                        'this' => [
+                                            'example' => "one",
+                                            'type' => 'string',
+                                        ],
+                                    ],
+                                ],
+                                [
+                                    'type' => 'object',
+                                    'description' => 'No, that one.',
+                                    'properties' => [
+                                        'that' => [
+                                            'example' => "one",
+                                            'type' => 'string',
+                                        ],
+                                    ],
+                                ],
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+        ], $results['paths']['/path1']['post']['responses']);
+    }
+
+    /** @test */
+    public function adds_more_than_two_answers_correctly_using_oneOf()
+    {
+        $endpointData1 = $this->createMockEndpointData([
+            'httpMethods' => ['POST'],
+            'uri' => '/path1',
+            'responses' => [
+                [
+                    'status' => 201,
+                    'description' => 'This one',
+                    'content' => '{"this": "one"}',
+                ],
+                [
+                    'status' => 201,
+                    'description' => 'No, that one.',
+                    'content' => '{"that": "one"}',
+                ],
+                [
+                    'status' => 201,
+                    'description' => 'No, another one.',
+                    'content' => '{"another": "one"}',
+                ],
+                [
+                    'status' => 200,
+                    'description' => 'A separate one',
+                    'content' => '{"the other": "one"}',
+                ],
+            ],
+        ]);
+        $groups = [$this->createGroup([$endpointData1])];
+
+        $results = $this->generate($groups);
+
+        $this->assertArraySubset([
+            '200' => [
+                'description' => 'A separate one',
+                'content' => [
+                    'application/json' => [
+                        'schema' => [
+                            'type' => 'object',
+                            'properties' => [
+                                'the other' => [
+                                    'example' => "one",
+                                    'type' => 'string',
+                                ],
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+            '201' => [
+                'description' => '',
+                'content' => [
+                    'application/json' => [
+                        'schema' => [
+                            'oneOf' => [
+                                [
+                                    'type' => 'object',
+                                    'description' => 'This one',
+                                    'properties' => [
+                                        'this' => [
+                                            'example' => "one",
+                                            'type' => 'string',
+                                        ],
+                                    ],
+                                ],
+                                [
+                                    'type' => 'object',
+                                    'description' => 'No, that one.',
+                                    'properties' => [
+                                        'that' => [
+                                            'example' => "one",
+                                            'type' => 'string',
+                                        ],
+                                    ],
+                                ],
+                                [
+                                    'type' => 'object',
+                                    'description' => 'No, another one.',
+                                    'properties' => [
+                                        'another' => [
+                                            'example' => "one",
+                                            'type' => 'string',
+                                        ],
+                                    ],
+                                ],
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+        ], $results['paths']['/path1']['post']['responses']);
+    }
+
+    /** @test */
+    public function adds_enum_values_to_response_properties()
+    {
+        $endpointData = $this->createMockEndpointData([
+            'httpMethods' => ['GEt'],
+            'uri' => '/path1',
+            'responses' => [
+                [
+                    'status' => 200,
+                    'description' => 'List of entities',
+                    'content' => '{"data":[{"name":"Resource name","uuid":"UUID","primary":true}]}',
+                ],
+            ],
+            'responseFields' => [
+                'data' => [
+                    'name' => 'data',
+                    'type' => 'array',
+                    'description' => 'Data wrapper',
+                ],
+                'data.name' => [
+                    'name' => 'Resource name',
+                    'type' => 'string',
+                    'description' => 'Name of the resource object',
+                    'required' => true,
+                ],
+                'data.uuid' => [
+                    'name' => 'Resource UUID',
+                    'type' => 'string',
+                    'description' => 'Unique ID for the resource',
+                    'required' => true,
+                ],
+                'data.primary' => [
+                    'name' => 'Is primary',
+                    'type' => 'bool',
+                    'description' => 'Is primary resource',
+                    'required' => true,
+                ],
+            ],
+        ]);
+
+        $groups = [$this->createGroup([$endpointData])];
+
+        $results = $this->generate($groups);
+
+        $this->assertArraySubset([
+            '200' => [
+                'description' => 'List of entities',
+                'content' => [
+                    'application/json' => [
+                        'schema' => [
+                            'type' => 'object',
+                            'properties' => [
+                                'data' => [
+                                    'type' => 'array',
+                                    'description' => 'Data wrapper',
+                                    'items' => [
+                                        'type' => 'object',
+                                        'properties' => [
+                                            'name' => [
+                                                'type' => 'string',
+                                                'description' => 'Name of the resource object',
+                                            ],
+                                            'uuid' => [
+                                                'type' => 'string',
+                                                'description' => 'Unique ID for the resource',
+                                            ],
+                                            'primary' => [
+                                                'type' => 'boolean',
+                                                'description' => 'Is primary resource',
+                                            ],
+                                        ],
+                                    ],
+                                    'required' => [
+                                        'name',
+                                        'uuid',
+                                        'primary',
+                                    ]
+                                ],
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+        ], $results['paths']['/path1']['get']['responses']);
+    }
+
+    /** @test */
+    public function lists_required_properties_in_request_body()
+    {
+        $endpointData = $this->createMockEndpointData([
+            'uri' => '/path',
+            'httpMethods' => ['POST'],
+            'bodyParameters' => [
+                'my_field' => [
+                    'name' => 'my_field',
+                    'description' => '',
+                    'required' => true,
+                    'example' => 'abc',
+                    'type' => 'string',
+                    'nullable' => false,
+                ],
+                'other_field.nested_field' => [
+                    'name' => 'nested_field',
+                    'description' => '',
+                    'required' => true,
+                    'example' => 'abc',
+                    'type' => 'string',
+                    'nullable' => false,
+                ],
+            ],
+        ]);
+        $groups = [$this->createGroup([$endpointData])];
+        $results = $this->generate($groups);
+
+        $this->assertArraySubset([
+            'requestBody' => [
+                'content' => [
+                    'application/json' => [
+                        'schema' => [
+                            'type' => 'object',
+                            'properties' => [
+                                'my_field' => [
+                                    'type' => 'string',
+                                ],
+                                'other_field' => [
+                                    'type' => 'object',
+                                    'properties' => [
+                                        'nested_field' => [
+                                            'type' => 'string',
+                                        ],
+                                    ],
+                                    'required' => ['nested_field'],
+                                ],
+                            ],
+                            'required' => ['my_field']
+                        ],
+                    ],
+                ],
+            ],
+        ], $results['paths']['/path']['post']);
+    }
+
+    /** @test */
+    public function can_extend_openapi_generator()
+    {
+        $endpointData1 = $this->createMockEndpointData([
+            'uri' => '/path',
+            'httpMethods' => ['POST'],
+            'custom' => ['permissions' => ['post:view']]
+        ]);
+        $groups = [$this->createGroup([$endpointData1])];
+        $extraGenerator = TestOpenApiGenerator::class;
+        $config = array_merge($this->config, [
+            'openapi' => [
+                'generators' => [
+                    $extraGenerator,
+                ],
+            ],
+        ]);
+        $writer = new OpenAPISpecWriter(new DocumentationConfig($config));
+
+        $results = $writer->generateSpecContent($groups);
+
+        $this->assertEquals([['default' => ['post:view']]], $results['paths']['/path']['post']['security']);
+    }
+
+    /** @test */
+    public function can_extend_openapi_generator_parameters()
+    {
+        $endpointData1 = $this->createMockEndpointData([
+            'uri' => '/{slug}/path',
+            'httpMethods' => ['POST'],
+            'custom' => ['permissions' => ['post:view']],
+            'urlParameters.slug' => [
+                'description' => 'Something',
+                'required' => true,
+                'example' => 56,
+                'type' => 'integer',
+                'name' => 'slug',
+            ],
+        ]);
+        $groups = [$this->createGroup([$endpointData1])];
+        $extraGenerator = ComponentsOpenApiGenerator::class;
+        $config = array_merge($this->config, [
+            'openapi' => [
+                'generators' => [
+                    $extraGenerator,
+                ],
+            ],
+        ]);
+        $writer = new OpenAPISpecWriter(new DocumentationConfig($config));
+
+        $results = $writer->generateSpecContent($groups);
+
+        $actualParameters = $results['paths']['/{slug}/path']['parameters'];
+        $this->assertCount(1, $actualParameters);
+        $this->assertEquals(['$ref' =>  "#/components/parameters/slugParam"], $actualParameters[0]);
+        $this->assertEquals([
+            'slugParam' => [
+                'in' => 'path',
+                'name' => 'slug',
+                'description' => 'The slug of the organization.',
+                'example' => 'acme-corp',
+                'required' => true,
+                'schema' => [
+                    'type' => 'string',
+                ],
+            ]
+        ], $results['components']['parameters']);
     }
 
     protected function createMockEndpointData(array $custom = []): OutputEndpointData
