@@ -3,7 +3,9 @@
 namespace Knuckles\Scribe\Tests\Unit;
 
 use Illuminate\Foundation\Application;
+use Illuminate\Support\Facades\Schema;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Translation\Translator;
 use Illuminate\Validation\Rule;
 use Illuminate\Validation\ValidationException;
 use Knuckles\Scribe\Extracting\ParsesValidationRules;
@@ -11,7 +13,6 @@ use Knuckles\Scribe\Tests\BaseLaravelTest;
 use Knuckles\Scribe\Tools\DocumentationConfig;
 use Knuckles\Scribe\Tests\Fixtures;
 
-$invokableRulesSupported = interface_exists(\Illuminate\Contracts\Validation\InvokableRule::class);
 $laravel10Rules = version_compare(Application::VERSION, '10.0', '>=');
 
 class ValidationRuleParsingTest extends BaseLaravelTest
@@ -39,6 +40,11 @@ class ValidationRuleParsingTest extends BaseLaravelTest
      */
     public function can_parse_supported_rules(array $ruleset, array $customInfo, array $expected)
     {
+        // Needed for `exists` rule
+        Schema::create('users', function ($table) {
+            $table->id();
+        });
+
         $results = $this->strategy->parse($ruleset, $customInfo);
 
         $parameterName = array_keys($ruleset)[0];
@@ -48,7 +54,9 @@ class ValidationRuleParsingTest extends BaseLaravelTest
             $this->assertEquals($expected['type'], $results[$parameterName]['type']);
         }
 
-        // Validate that the generated values actually pass validation
+        // Validate that the generated values actually pass validation (for rules where we can generate some data)
+        if (is_string($ruleset[$parameterName]) && str_contains($ruleset[$parameterName], "exists")) return;
+
         $exampleData = [$parameterName => $results[$parameterName]['example']];
         $validator = Validator::make($exampleData, $ruleset);
         try {
@@ -65,14 +73,13 @@ class ValidationRuleParsingTest extends BaseLaravelTest
     public function can_parse_rule_objects()
     {
         $results = $this->strategy->parse([
-            'in_param' => ['numeric', Rule::in([3,5,6])]
+            'in_param' => ['numeric', Rule::in([3, 5, 6])]
         ]);
         $this->assertEquals(
-            'Must be one of <code>3</code>, <code>5</code>, or <code>6</code>.',
-            $results['in_param']['description']
+            [3, 5, 6],
+            $results['in_param']['enumValues']
         );
     }
-
 
     /** @test */
     public function can_transform_arrays_and_objects()
@@ -110,6 +117,17 @@ class ValidationRuleParsingTest extends BaseLaravelTest
         $this->assertEquals('string[]', $results['array_of_objects_with_array[].another[].one.field1']['type']);
         $this->assertEquals('integer', $results['array_of_objects_with_array[].another[].one.field2']['type']);
         $this->assertEquals('number', $results['array_of_objects_with_array[].another[].two.field2']['type']);
+
+        $ruleset = [
+            '*.foo' => 'required|array',
+            '*.foo.*' => 'required|array',
+            '*.foo.*.bar' => 'required',
+        ];
+        $results = $this->strategy->parse($ruleset);
+        $this->assertCount(3, $results);
+        $this->assertEquals('object', $results['*']['type']);
+        $this->assertEquals('object[]', $results['*.foo']['type']);
+        $this->assertEquals('string', $results['*.foo[].bar']['type']);
     }
 
     public static function supportedRules()
@@ -228,8 +246,9 @@ class ValidationRuleParsingTest extends BaseLaravelTest
             ['in_param' => 'in:3,5,6'],
             ['in_param' => ['description' => $description]],
             [
-                'description' => "$description. Must be one of <code>3</code>, <code>5</code>, or <code>6</code>.",
+                'description' => $description . ".",
                 'type' => 'string',
+                'enumValues' => [3, 5, 6]
             ],
         ];
         yield 'not_in' => [
@@ -383,6 +402,11 @@ class ValidationRuleParsingTest extends BaseLaravelTest
             [],
             ['description' => "Must be a file. Must not be greater than 6 kilobytes."],
         ];
+        yield 'max (untyped)' => [
+            ['max_param' => 'max:6'],
+            [],
+            ['description' => "Must not be greater than 6 characters."],
+        ];
         yield 'min (number)' => [
             ['min_param' => 'numeric|min:6'],
             [],
@@ -426,21 +450,26 @@ class ValidationRuleParsingTest extends BaseLaravelTest
                 'description' => 'Must be accepted.',
             ],
         ];
+        yield 'exists' => [
+            ['exists_param' => 'exists:users,id'],
+            [],
+            [
+                'description' => 'The <code>id</code> of an existing record in the users table.',
+            ],
+        ];
         yield 'unsupported' => [
             ['unsupported_param' => [new DummyValidationRule, 'bail']],
             ['unsupported_param' => ['description' => $description]],
             ['description' => "$description."],
         ];
-        if (version_compare(Application::VERSION, '8.53', '>=')) {
-            yield 'accepted_if' => [
-                ['accepted_if_param' => 'accepted_if:another_field,a_value'],
-                [],
-                [
-                    'type' => 'boolean',
-                    'description' => "Must be accepted when <code>another_field</code> is <code>a_value</code>.",
-                ],
-            ];
-        }
+        yield 'accepted_if' => [
+            ['accepted_if_param' => 'accepted_if:another_field,a_value'],
+            [],
+            [
+                'type' => 'boolean',
+                'description' => "Must be accepted when <code>another_field</code> is <code>a_value</code>.",
+            ],
+        ];
     }
 
     /** @test */
@@ -505,10 +534,7 @@ class ValidationRuleParsingTest extends BaseLaravelTest
             'param1' => ['bail', 'required', new DummyWithDocsValidationRule],
         ];
 
-        global $invokableRulesSupported;
-        if ($invokableRulesSupported) {
-            $ruleset['param2'] = [new DummyInvokableValidationRule];
-        }
+        $ruleset['param2'] = [new DummyInvokableValidationRule];
         global $laravel10Rules;
         if ($laravel10Rules) {
             $ruleset['param3'] = [new DummyL10ValidationRule];
@@ -517,58 +543,160 @@ class ValidationRuleParsingTest extends BaseLaravelTest
         $results = $this->strategy->parse($ruleset);
         $this->assertEquals(true, $results['param1']['required']);
         $this->assertEquals('This is a dummy test rule.', $results['param1']['description']);
-        if (isset($results['param2'])) $this->assertEquals('This rule is invokable.', $results['param2']['description']);
-       if (isset($results['param3'])) $this->assertEquals('This is a custom rule.', $results['param3']['description']);
+        $this->assertEquals('This rule is invokable.', $results['param2']['description']);
+        if (isset($results['param3'])) $this->assertEquals('This is a custom rule.', $results['param3']['description']);
     }
 
     /** @test */
     public function can_parse_enum_rules()
     {
-        if (phpversion() < 8.1) {
-            $this->markTestSkipped('Enums are only supported in PHP 8.1 or later');
-        }
-
         $results = $this->strategy->parse([
-            'enum' => ['required', Rule::enum(Fixtures\TestStringBackedEnum::class)],
+            'enum' => [
+                'required',
+                Rule::enum(Fixtures\TestStringBackedEnum::class)
+            ],
         ]);
         $this->assertEquals('string', $results['enum']['type']);
         $this->assertEquals(
-            'Must be one of <code>red</code>, <code>green</code>, or <code>blue</code>.',
-            $results['enum']['description']
+            ['red', 'green', 'blue'],
+            $results['enum']['enumValues']
         );
         $this->assertTrue(in_array(
             $results['enum']['example'],
-            array_map(fn ($case) => $case->value, Fixtures\TestStringBackedEnum::cases())
+            array_map(fn($case) => $case->value, Fixtures\TestStringBackedEnum::cases())
         ));
 
 
         $results = $this->strategy->parse([
-            'enum' => ['required', Rule::enum(Fixtures\TestIntegerBackedEnum::class)],
+            'enum' => [
+                'required',
+                new \Illuminate\Validation\Rules\Enum(Fixtures\TestIntegerBackedEnum::class),
+                // Not supported in Laravel 8
+                // Rule::enum(Fixtures\TestIntegerBackedEnum::class)
+            ],
         ]);
         $this->assertEquals('integer', $results['enum']['type']);
         $this->assertEquals(
-            'Must be one of <code>1</code>, <code>2</code>, or <code>3</code>.',
-            $results['enum']['description']
+            [1, 2, 3],
+            $results['enum']['enumValues']
         );
         $this->assertTrue(in_array(
             $results['enum']['example'],
-            array_map(fn ($case) => $case->value, Fixtures\TestIntegerBackedEnum::cases())
+            array_map(fn($case) => $case->value, Fixtures\TestIntegerBackedEnum::cases())
         ));
 
         $results = $this->strategy->parse([
-            'enum' => ['required', Rule::enum(Fixtures\TestStringBackedEnum::class)],
+            'enum' => [
+                'required',
+                new \Illuminate\Validation\Rules\Enum(Fixtures\TestStringBackedEnum::class),
+                // Not supported in Laravel 8
+                // Rule::enum(Fixtures\TestStringBackedEnum::class),
+            ],
         ], [
             'enum' => ['description' => 'A description'],
         ]);
         $this->assertEquals('string', $results['enum']['type']);
         $this->assertEquals(
-            'A description. Must be one of <code>red</code>, <code>green</code>, or <code>blue</code>.',
+            'A description.',
             $results['enum']['description']
         );
         $this->assertTrue(in_array(
             $results['enum']['example'],
-            array_map(fn ($case) => $case->value, Fixtures\TestStringBackedEnum::cases())
+            array_map(fn($case) => $case->value, Fixtures\TestStringBackedEnum::cases())
         ));
+    }
+
+    /** @test */
+    public function can_translate_validation_rules_with_types_with_translator_without_array_support()
+    {
+        // Single line DocComment
+        $ruleset = [
+            'nested' => [
+                'string', 'max:20',
+            ],
+        ];
+
+        $results = $this->strategy->parse($ruleset);
+
+        $this->assertEquals('Must not be greater than 20 characters.', $results['nested']['description']);
+
+        $this->app->extend('translator', function ($command, $app) {
+            $loader = $app['translation.loader'];
+            $locale = $app['config']['app.locale'];
+            return new DummyTranslator($loader, $locale);
+        });
+
+        $results = $this->strategy->parse($ruleset);
+
+        $this->assertEquals('successfully translated by concatenated string.', $results['nested']['description']);
+
+    }
+
+    /** @test */
+    public function can_parse_nullable_rules()
+    {
+        $ruleset = [
+            'nullable_param' => 'nullable|string',
+        ];
+
+        $results = $this->strategy->parse($ruleset);
+
+        $this->assertTrue($results['nullable_param']['nullable']);
+
+        $ruleset = [
+            'nullable_param' => 'string',
+        ];
+
+        $results = $this->strategy->parse($ruleset);
+
+        $this->assertFalse($results['nullable_param']['nullable']);
+
+        $ruleset = [
+            'required_param' => 'required|nullable|string',
+        ];
+
+        $results = $this->strategy->parse($ruleset);
+
+        $this->assertFalse($results['required_param']['nullable']);
+
+
+        $ruleset = [
+            'array_param' => 'array',
+            'array_param.*.field' => 'nullable|string',
+        ];
+
+        $results = $this->strategy->parse($ruleset);
+
+        $this->assertFalse($results['array_param']['nullable']);
+        $this->assertTrue($results['array_param[].field']['nullable']);
+
+        $ruleset = [
+            'object' => 'array',
+            'object.field1' => 'string',
+            'object.field2' => 'nullable|string',
+        ];
+
+        $results = $this->strategy->parse($ruleset);
+
+        $this->assertFalse($results['object']['nullable']);
+        $this->assertFalse($results['object.field1']['nullable']);
+        $this->assertTrue($results['object.field2']['nullable']);
+    }
+
+    /** @test */
+    public function can_parse_rules_which_reference_other_fields()
+    {
+        $ruleset = [
+            'to_time' => 'date|max:6',
+            'from_time' => [
+                'date',
+                'before:to_time',
+            ],
+        ];
+
+        $results = $this->strategy->parse($ruleset);
+
+        $this->assertEquals("Must be a valid date. Must be a date before <code>to_time</code>.", $results['from_time']['description']);
     }
 }
 
@@ -606,28 +734,27 @@ class DummyWithDocsValidationRule implements \Illuminate\Contracts\Validation\Ru
     }
 }
 
-if ($invokableRulesSupported) {
-    class DummyInvokableValidationRule implements \Illuminate\Contracts\Validation\InvokableRule
+// Laravel 9 introduced InvokableRule
+class DummyInvokableValidationRule implements \Illuminate\Contracts\Validation\InvokableRule
+{
+    public function __invoke($attribute, $value, $fail)
     {
-        public function __invoke($attribute, $value, $fail)
-        {
-            if (strtoupper($value) !== $value) {
-                $fail(':attribute must be uppercase.');
-            }
+        if (strtoupper($value) !== $value) {
+            $fail(':attribute must be uppercase.');
         }
+    }
 
-        public function docs()
-        {
+    public function docs()
+    {
 
-            return [
-                'description' => 'This rule is invokable.',
-            ];
-        }
+        return [
+            'description' => 'This rule is invokable.',
+        ];
     }
 }
 
 if ($laravel10Rules) {
-// Laravel 10 deprecated the previous Rule and InvokableRule classes for a single interface
+    // Laravel 10 deprecated the previous Rule and InvokableRule classes for a single interface
     // (https://github.com/laravel/framework/pull/45954)
     class DummyL10ValidationRule implements \Illuminate\Contracts\Validation\ValidationRule
     {
@@ -644,5 +771,17 @@ if ($laravel10Rules) {
                 'description' => 'This is a custom rule.',
             ];
         }
+    }
+}
+
+class DummyTranslator extends Translator
+{
+    public function get($key, array $replace = [], $locale = null, $fallback = true)
+    {
+        if ($key === 'validation.max.string') {
+            return 'successfully translated by concatenated string';
+        }
+
+        return $key;
     }
 }
